### 2021-03-23 #3129 @danielskatz
👋 @nkallima - I'll be your editor.

1. Can you suggest any potential reviewers by mentioning them here in this thread (without tagging them with an @)? [This list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
2. Also, can you help me understand the authorship? It looks like you are the only author, but I see multiple contributors in the  repository.
