### 2021-03-24 #3129 @nkallima
@danielskatz – Thank you for handling the paper.
I checked the list of the reviewers and judging from the domains/topic areas, a list of potential reviewers could be the following:
-	Kanirudh
-	Jpenuchot
-	jarrah42
-	cedricchevalier19

Most of the content on this repository is an extension of my research on concurrent data-structures/algorithms and performance evaluation performed in the last decade. Recently, my colleague Spiros Agathos used the software extensively and gently contributed a few patches/fixes. Moreover, Eftychia Datsika (edatsika) improved the logo of the repository. I acknowledge Spiros and Eftychia in the Acknowledgment section of the paper.
