### 2021-03-24 #3129 @danielskatz
👋 @Jpenuchot - would you be able to review this submission for JOSS?
