### 2021-03-26 #3129 @danielskatz
👋 @jarrah42 - would you be able to review this submission for JOSS?
