### 2021-03-30 #3129 @danielskatz
👋 @mkaguilera - Would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews of the software and paper here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
