### 2021-04-01 ef0f56a
Adding LCRQ (Morrison and Afek, PPoPP 2013, http://mcg.cs.tau.ac.il/projects/lcrq) to the framework.
