+++ validate.sh
@@ -33,7 +33,7 @@ COLOR_PASS="[ \e[32mPASS\e[39m ]"
 COLOR_FAIL="[ \e[31mFAIL\e[39m ]"
 
 declare -a uobjects=("ccsynchbench.run"                     "dsmsynchbench.run" "hsynchbench.run" "oscibench.run"      "simbench.run")
-declare -a queues=(  "ccqueuebench.run" "clhqueuebench.run" "dsmqueuebench.run" "hqueuebench.run" "osciqueuebench.run" "simqueuebench.run")
+declare -a queues=(  "ccqueuebench.run" "clhqueuebench.run" "dsmqueuebench.run" "hqueuebench.run" "osciqueuebench.run" "simqueuebench.run" "lcrqbench.run")
 declare -a stacks=(  "ccstackbench.run" "clhstackbench.run" "dsmstackbench.run" "hstackbench.run" "oscistackbench.run" "simstackbench.run")
 
 if [ "$1" = "--help" ] || [ "$1" = "-h" ]; then
