+++ libconcurrent/primitives/primitives.c
@@ -12,39 +12,43 @@ extern __thread int64_t __executed_faa;
 #endif
 
 #ifdef __OLD_GCC_X86__
-inline static bool __CASPTR(void *A, void *B, void *C) {
+inline bool __CASPTR(void *A, void *B, void *C) {
     uint64_t prev;
     uint64_t *p = (uint64_t *)A;
 
-    asm volatile("lock;cmpxchgq %1,%2" 
+    asm volatile("lock;"
+                 "cmpxchgq %1,%2" 
                  : "=a"(prev)
                  : "r"((uint64_t)C), "m"(*p), "0"((uint64_t)B) 
                  : "memory");
     return (prev == (uint64_t)B);
 }
 
-inline static bool __CAS64(volatile uint64_t *A, uint64_t B, uint64_t C) {
+inline bool __CAS64(volatile uint64_t *A, uint64_t B, uint64_t C) {
     uint64_t prev;
     uint64_t *p = (uint64_t *)A;
 
-    asm volatile("lock;cmpxchgq %1,%2"
+    asm volatile("lock;"
+                 "cmpxchgq %1,%2"
                  : "=a"(prev)
                  : "r"(C), "m"(*p), "0"(B)
                  : "memory");
     return (prev == B);
 }
 
-inline static bool __CAS32(uint32_t *A, uint32_t B, uint32_t C) {
+inline bool __CAS32(uint32_t *A, uint32_t B, uint32_t C) {
     uint32_t prev;
     uint32_t *p = (uint32_t *)A;
 
-    asm volatile("lock;cmpxchgl %1,%2" : "=a"(prev) 
+    asm volatile("lock;"
+                 "cmpxchgl %1,%2" 
+                 : "=a"(prev) 
                  : "r"(C), "m"(*p), "0"(B) 
                  : "memory");
     return (prev == B);
 }
 
-inline static void *__SWAP(void *A, void *B) {
+inline void *__SWAP(void *A, void *B) {
     int64_t *p = (int64_t *)A;
 
     asm volatile("lock;"
@@ -55,7 +59,7 @@ inline static void *__SWAP(void *A, void *B) {
     return B;
 }
 
-inline static int64_t __FAA64(volatile int64_t *A, int64_t B) {
+inline int64_t __FAA64(volatile int64_t *A, int64_t B) {
     asm volatile("lock;"
                  "xaddq %0, %1"
                  : "=r"(B), "=m"(*A)
@@ -64,13 +68,25 @@ inline static int64_t __FAA64(volatile int64_t *A, int64_t B) {
     return B;
 }
 
-inline static int32_t __FAA32(volatile int32_t *A, int32_t B) {
+inline int32_t __FAA32(volatile int32_t *A, int32_t B) {
     asm volatile("lock;"
                  "xaddl %0, %1"
                  : "=r"(B), "=m"(*A)
                  : "0"(B), "m"(*A)
                  : "memory");
     return B;
+}
+
+inline uint64_t __BitTAS64(volatile uint64_t *A, unsigned char B) {
+    int64_t *p = (int64_t *)A;
+    int64_t bit = B;
+    asm volatile("lock;" 
+                 "btsq %0, %1"
+                 : "=r"(bit), "=m"(*p)
+                 : "0"(bit), "m"(*p)
+                 : "memory");
+
+    return bit;
 }
 
 inline int bitSearchFirst(uint64_t B) {
@@ -90,6 +106,30 @@ inline uint64_t nonZeroBits(uint64_t v) {
     return c;
 }
[…]
