### 2021-04-01 #3143 @whedon
[REVIEW]: Synch: A framework for concurrent data-structures and benchmarks
**Submitting author:** @nkallima (<a href="http://orcid.org/0000-0002-0331-1475">Nikolaos D. Kallimanis</a>)
**Repository:** <a href="https://github.com/nkallima/sim-universal-construction" target ="_blank">https://github.com/nkallima/sim-universal-construction</a>
**Version:** v3.0.0

- [x] **State of the field:** Do the authors describe how this software compares to other commonly-used packages? Mentioned in the background section
