### 2021-04-01 #3143 @danielskatz
👋 @nkallima - note that there a couple of references in your bib file that you can fix (by removing the `https://doi.org/` part) when you get a chance - no hurry though...  You can use the same `@whedon` comments I did above to recheck and regenerate the pdf, too.
