+++ paper.bib
@@ -37,7 +37,7 @@
   title={Highly-efficient wait-free synchronization},
   author={Fatourou, Panagiota and Kallimanis, Nikolaos D},
   journal={Theory of Computing Systems},
-  doi={https://doi.org/10.1007/s00224-013-9491-y},
+  doi={10.1007/s00224-013-9491-y},
   volume={55},
   number={3},
   pages={475--520},
@@ -141,7 +141,7 @@
   address="Berlin, Heidelberg",
   pages="650--661",
   isbn="978-3-642-32820-6",
-  doi={https://doi.org/10.1007/978-3-642-32820-6_64}
+  doi={10.1007/978-3-642-32820-6_64}
 }
 
 @inproceedings{MS96,
