+++ paper.bib
@@ -87,8 +87,7 @@
   title={Executing parallel programs with synchronization bottlenecks efficiently},
   author={Oyama, Yoshihiro and Taura, Kenjiro and Yonezawa, Akinori},
   booktitle={Proceedings of the International Workshop on Parallel and Distributed Computing for Symbolic and Irregular Applications},
-  URL="https://ci.nii.ac.jp/naid/110002725744/en/",
-  doi={10.1.1.159.183},
+  doi={10.1142/4278},
   volume={16},
   pages={95},
   year={1999},
