### 2021-04-06 357f7ed
Introducing LCRQ in the v2.4.0 development version. (#26)
Adding LCRQ (Morrison and Afek, PPoPP 2013, http://mcg.cs.tau.ac.il/projects/lcrq) to the framework.
