### 2021-04-13 5ec77a6
Support for memory reclamation in SimStack. (#27)
Memory reclamation for SimStack is now supported.
