### 2021-04-13 85c9128
Support for memory reclamation in SimStack. (#27)
Memory reclamation for SimStack is now supported.
