### 2021-04-13 #3143 @jarrah42
I see two other contributors on the project, but they are not listed in the paper.
