### 2021-04-14 #3143 @nkallima
@jarrah42

>
>
> I see two other contributors on the project, but they are not listed in the paper.

First of all, thank you for your feedback. Most of the content on this repository is an extension of my research on concurrent data-structures/algorithms and performance evaluation. Recently, my colleague Spiros Agathos used the software extensively and gently contributed a few patches/fixes. Moreover, Eftychia Datsika (edatsika) improved the logo of the repository. I acknowledge Spiros and Eftychia in the Acknowledgment section of the paper.

>
>
> The project is missing installation instructions, user/developer documentation, and API documentation.

In the Readme file, you can find information on how to build the framework, and how to execute any of the provided benchmarks. Currently, no special installer is provided, since most of the users (i.e., researchers/developers of concurrent algorithms) heavily modify the provided data-structures and benchmarks in order to develop new algorithms. Moreover, most of the users use it on large multi-processors without administrative rights and thus an installer would not be very valuable.
In the Readme file you can find a very simple example of the API usage. However, I understand that this is far from complete. Thus, I could provide some better documentation for the basic components but I would need some time for this. Do you think that this is OK?
