### 2021-04-14 #3143 @jarrah42
> > The project is missing installation instructions, user/developer documentation, and API documentation.
>
> In the Readme file, you can find information on how to build the framework, and how to execute any of the provided benchmarks. Currently, no special installer is provided, since most of the users (i.e., researchers/developers of concurrent algorithms) heavily modify the provided data-structures and benchmarks in order to develop new algorithms. Moreover, most of the users use it on large multi-processors without administrative rights and thus an installer would not be very valuable.
> In the Readme file you can find a very simple example of the API usage. However, I understand that this is far from complete. Thus, I could provide some better documentation for the basic components but I would need some time for this. Do you think that this is OK?

I think we can work with you until the review criteria is satisfied. It's not like a paper submission where you only get on chance.

Regarding the Readme. I would like to see a little more detail on how to build the software. E.g. where do you run the `make` command? Are there other configuration steps? Most packages these days provide some kind on installation command that will at least copy the header files and libraries to a specified location. Using CMake or similar will do this for you automatically.

I would definitely like to see complete API documentation, and better documentation on how to use the library.

Thanks!
