### 2021-04-18 7145832
Documentation for CCSynch using doxygen. Cleanup of includes.
