+++ README.md
@@ -6,18 +6,70 @@
 
 # Summary
 
-This is an open-source framework for concurrent data-structures and benchmarks. The provided framework 
-contains a substantial set of concurrent data-structures such as `queues`, `stacks`, `combining-objects`,
-`hash-tables`, `locks`, etc. This framework also provides a user-friendly runtime for developing and
-benchmarking concurrent data-structures. Among other features, this runtime provides functionality 
-for creating threads easily (both Posix and user-level threads), tools for measuring performance, etc.
-The provided concurrent data-structures and the runtime are highly optimized for contemporary 
-NUMA multiprocessors such as AMD Epyc and Intel Xeon.
+This is an open-source framework for concurrent data-structures and benchmarks. The provided framework contains a substantial set of concurrent data-structures such as `queues`, `stacks`, `combining-objects`,
+`hash-tables`, `locks`, etc. This framework also provides a user-friendly runtime for developing and benchmarking concurrent data-structures. Among other features, this runtime provides functionality for creating threads easily (both Posix and user-level threads), tools for measuring performance, etc. The provided concurrent data-structures and the runtime are highly optimized for contemporary NUMA multiprocessors such as AMD Epyc and Intel Xeon.
 
-The current version of this code is optimized for x86_64 machine architecture, but the code is also
-successfully tested in other machine architectures, such as ARM-V8 and RISC-V. 
-Some of the benchmarks perform much better in architectures that natively support Fetch&Add
-instructions (e.g., x86_64, etc.).
+The current version of this code is optimized for x86_64 machine architecture, but the code is also successfully tested in other machine architectures, such as ARM-V8 and RISC-V. Some of the benchmarks perform much better in architectures that natively support Fetch&Add instructions (e.g., x86_64, etc.).
+
+
+# Collection
+
+The current version of this library provides the following concurrent data-structures implementations:
+
+| Concurrent  Object    |                Provided Implementations                           |
+| --------------------- | ----------------------------------------------------------------- |
+| Combining Objects     | CC-Synch, DSM-Synch and H-Synch [1]                               |
+|                       | PSim [2,10]                                                       |
+|                       | Osci [3]                                                          |
+|                       | Oyama [4]                                                         |
+| Concurrent Queues     | CC-Queue, DSM-Queue and H-Queue [1]                               |
+|                       | SimQueue [2,10]                                                   |
+|                       | OsciQueue [3]                                                     |
+|                       | CLH-Queue [5,6]                                                   |
+|                       | MS-Queue [7]                                                      |
+|                       | LCRQ [11,12]                                                      |
+| Concurrent Stacks     | CC-Stack, DSM-Stack and H-Stack [1]                               |
+|                       | SimStack [2,10]                                                   |
+|                       | OsciStack [3]                                                     |
+|                       | CLH-Stack [5,6]                                                   |
+|                       | LF-Stack [8]                                                      |
+| Locks                 | CLH [5,6]                                                         |
+|                       | MCS [9]                                                           |
+| Hash Tables           | CLH-Hash [5,6]                                                    |
+|                       | A hash-table based on DSM-Synch [1]                               |
+
+
+# Requirements
+
+- A modern 64-bit machine. Currently, 32-bit architectures are not supported. The current version of this code is optimized for the x86_64 machine architecture, but the code is also successfully tested in other machine architectures, such as ARM-V8 and RISC-V. Some of the benchmarks perform much better in architectures that natively support Fetch&Add instructions (e.g., x86_64, etc.).
+- As a compiler, gcc of version 4.8 or greater is recommended, but you may also try to use icx or clang. 
[…]
