### 2021-04-22 e1e0f74
Adding the ability for installing the framework, i.e. `make install`.
