+++ README.md
@@ -53,30 +53,36 @@ The following table presents a summary of the concurrent data-structures offered
     - `libatomic`
     - `libnuma`
     - `libpapi` in case that the `_TRACK_CPU_COUNTERS` flag is enabled in `libconcurrent/config.h`.
+- For building the documentation (i.e. man-pages), `doxygen` is required.
 
-For getting the best performance, some modifications in Makefiles may be needed (compiler flags, etc.). Important parameters for the benchmarks and/or library are placed in the `config.h` file (see more on Performance/Optimizations Section).
 
-
-# Configuring and compiling the framework
-
-In case that you just want to compile the library that provides all the implemented concurrent algorithms execute one of the following make commands in the root directory of the source files. This step is necessary in case that you want to run benchmarks.
-
-|     Command             |                                Description                                                                    |
-| ----------------------- | ------------------------------------------------------------------------------------------------------------- |
-|  `make`                 |  Auto-detects the current architecture and compiles the source-code for it (this should work for most users). |
-|  `make CC=cc ARCH=arch` |  Compiles the source-code for the current architecture using the cc compiler.                                 |
-|  `make clang`           |  Compiles the source-code using the clang compiler.                                                           |
-|  `make icx`             |  Compiles the source-code using the Intel icx compiler.                                                       |
-|  `make unknown`         |  Compiles the source-code for architectures other than X86_64, e.g. RISC-V, ARM, etc.                         |
-|  `make clean`           |  Cleaning-up all the binary files.                                                                            |
-|  `make docs`            |  Creating the documentation (i.e. man-pages for the framework).                                               |
+# Configuring, compiling and installing the framework
 
 In the `libconcurrent/config.h` file, the user can configure some basic options for the framework, such as:
 - Enable/disable debug mode.
 - Support for Numa machines.
 - Enable performance statistics, etc.
 
-However, the default configuration should work well in many cases.
+The provided default configuration should work well in many cases. However, the default configuration may not provide the best performance. For getting the best performance, modifying the `libconcurrent/config.h` may be needed (see more on Performance/Optimizations Section).
+
+In case that you want to compile the library that provides all the implemented concurrent algorithms just execute `make` in the root directory of the source files. This step is necessary in case that you want to run benchmarks. However, some extra make options are provided in case the user wants to compile the framework with other than system's default compiler, clean the binary files, etc. The following table provides the list with all the available make options.
+
+|     Command             |                                Description                                                                    |
+| ----------------------- | ------------------------------------------------------------------------------------------------------------- |
+|  `make`                 |  Auto-detects the current architecture and compiles the source-code for it (this should work for most users). |
+|  `make CC=cc ARCH=arch` |  Compiles the source-code for the current architecture using the `cc` compiler.                               |
+|  `make clang`           |  Compiles the source-code using the clang compiler.                                                           |
+|  `make icx`             |  Compiles the source-code using the Intel icx compiler.                                                       |
+|  `make unknown`         |  Compiles the source-code for architectures other than X86_64, e.g. RISC-V, ARM, etc.                         |
+|  `make clean`           |  Cleaning-up all the binary files.                                                                            |
[…]
