### 2021-04-22 38d4447
Fixining installation instructions on README.md.
