+++ bench.sh
@@ -32,6 +32,7 @@ RUNS=""
 LIST=0
 WORKLOAD="-w 64"
 NUMA_NODES=""
+SCRIPTPATH=$(dirname ${BASH_SOURCE[0]})
 
 if [ "$#" = "0" ] || [ "$1" = "--help" ] || [ "$1" = "-h" ]; then
     usage;
@@ -90,15 +91,15 @@ while [ "$1" != "" ]; do
 done
 
 if [ $LIST = "1" ]; then
-   cd build/bin;
+   cd $SCRIPTPATH/build/bin;
    ls -lafr *.run
    exit -1;
 fi
 
-if [ ! -e build/bin/$FILE ]; then
+if [ ! -e $SCRIPTPATH/build/bin/$FILE ]; then
    echo -e "\n" $FILE "is not available for benchmarking.\n"
    echo -e "Available files for benchmarking: "
-   cd build/bin;
+   cd $SCRIPTPATH/build/bin;
    ls -lafr *.run
    exit -1;
 fi
@@ -141,7 +142,7 @@ for PTHREADS in "${PTHREADS_ARRAY[@]}"; do
     
     # Redirect stdout to res.txt, stderr to /dev/null
     for (( i=1; i<=$ITERATIONS; i++ ));do
-        ./build/bin/$FILE -t $PTHREADS $WORKLOAD $FIBERS $RUNS $NUMA_NODES $BACKOFF $MIN_BACKOFF 1>> res.txt 2> /dev/null;
+        $SCRIPTPATH/build/bin/$FILE -t $PTHREADS $WORKLOAD $FIBERS $RUNS $NUMA_NODES $BACKOFF $MIN_BACKOFF 1>> res.txt 2> /dev/null;
     done
 
     awk 'BEGIN {debug_prefix="";
