### 2021-05-01 6a6364b
Removing SYNCH_COMPACT_ALLOCATION from the code of CCSynch and DSMSynch. SYNCH_COMPACT_ALLOCATION is now replaced with the getMachineModel() functionality that returns the processor's vendor (e.g. AMD, Intel, etc.). However, SYNCH_COMPACT_ALLOCATION is still in use (at least for now) in system.h for appropriately configuring the size of CACHE_LINE.
