+++ libconcurrent/primitives/primitives.c
@@ -8,6 +8,8 @@
 #    include <numa.h>
 #endif
 
+#define MAX_VENDOR_STR_SIZE 64
+
 #ifdef DEBUG
 extern __thread int64_t __failed_cas;
 extern __thread int64_t __executed_cas;
@@ -185,6 +187,36 @@ inline int64_t getTimeMillis(void) {
         perror("clock_gettime");
         return 0;
     } else return tm.tv_sec*1000LL + tm.tv_nsec/1000000LL;
+}
+
+inline uint64_t getMachineModel(void) {
+#if defined(__amd64__) || defined(__x86_64__)
+    char cpu_model[MAX_VENDOR_STR_SIZE] = {'\0'};
+
+    asm volatile("movl $0, %%eax\n"
+                 "cpuid\n"
+                 "movl %%ebx, %0\n"
+                 "movl %%edx, %1\n"
+                 "movl %%ecx, %2\n"
+                 : "=m"(cpu_model[0]), "=m"(cpu_model[4]), "=m"(cpu_model[8])
+                 :: "%eax", "%ebx", "%edx", "%ecx", "memory");
+#ifdef DEBUG
+    fprintf(stderr, "DEBUG: Machine model: %s\n", cpu_model);
+#endif
+
+    if (strcmp(cpu_model, "AuthenticAMD") == 0)
+        return AMD_X86_MACHINE;
+    else if (strcmp(cpu_model, "GenuineIntel") == 0)
+        return INTEL_X86_MACHINE;
+    else
+        return X86_GENERIC_MACHINE;
+#elif defined(__aarch64__)
+    return ARM_GENERIC_MACHINE;
+#elif defined(__riscv__) || defined(__riscv)
+    return RISCV_GENERIC_MACHINE;
+#else
+    return UNKNOWN_MACHINE;
+#endif
 }
 
 inline bool _CASPTR(void *A, void *B, void *C) {
