+++ README.md
@@ -70,7 +70,7 @@ In case that you want to compile the library that provides all the implemented c
 |     Command             |                                Description                                                                    |
 | ----------------------- | ------------------------------------------------------------------------------------------------------------- |
 |  `make`                 |  Auto-detects the current architecture and compiles the source-code for it (this should work for most users). |
-|  `make CC=cc ARCH=arch` |  Compiles the source-code for the current architecture using the `cc` compiler.                               |
+|  `make CC=cc`           |  Compiles the source-code for the current architecture using the `cc` compiler.                               |
 |  `make clang`           |  Compiles the source-code using the clang compiler.                                                           |
 |  `make icx`             |  Compiles the source-code using the Intel icx compiler.                                                       |
 |  `make unknown`         |  Compiles the source-code for architectures other than X86_64, e.g. RISC-V, ARM, etc.                         |
@@ -107,7 +107,7 @@ The following options are available:
 |  `-bl`, `--backoff_low` |  set a lower backoff bound (only for msqueue, lfstack and lfuobject benchmarks)                                                                  |
 |  `-h`, `--help`         |  displays this help and exits                                                                                                                    |
 
-The framework provides the `validate.sh` validation/smoke script. The `validate.sh` script compiles the sources in `DEBUG` mode and runs a big set of benchmarks with various numbers of threads. After running each of the benchmarks, the script evaluates the `DEBUG` output and in case of success it prints `PASS`. In case of a failure, the script simply prints `FAIL`. In order to see all the available options of the validation/smoke script, execute `validate.sh -h`. The following image shows the execution and the default behavior of `validate.sh`.
+The framework provides the `validate.sh` validation/smoke script. The `validate.sh` script compiles the sources in `DEBUG` mode and runs a big set of benchmarks with various numbers of threads. After running each of the benchmarks, the script evaluates the `DEBUG` output and in case of success it prints `PASS`. In case of a failure, the script simply prints `FAIL`. In order to see all the available options of the validation/smoke script, execute `validate.sh -h`. Given that the `validate.sh` validation/smoke script depends on binaries that are compiled in `DEBUG` mode, it is not installed while executing `make install`. The following image shows the execution and the default behavior of `validate.sh`.
 
 ![](resources/validate_example.gif)
 
