### 2021-05-01 e40278a
Documentation for CC-Synch, DSM-Synch and H-Synch combining objects.
