+++ libconcurrent/includes/hsynch.h
@@ -1,3 +1,12 @@
+/// @file hsynch.h
+/// @author Nikolaos D. Kallimanis (nkallima@gmail.com)
+/// @brief This file exposes the API of the HSynch combining object.
+/// An example of use of this API is provided in benchmarks/hsynchbench.c file.
+///
+/// For a more detailed description see the original publication:
+/// Panagiota Fatourou, and Nikolaos D. Kallimanis."Revisiting the combining synchronization technique".
+/// ACM SIGPLAN Notices. Vol. 47. No. 8. ACM, PPoPP 2012.
+/// @copyright Copyright (c) 2021
 #ifndef _HSYNCH_H_
 #define _HSYNCH_H_
 
@@ -5,50 +14,107 @@
 #include <primitives.h>
 #include <clh.h>
 
+/// @brief Whenever numa_regions is equal to HSYNCH_DEFAULT_NUMA_POLICY, the user uses the default number of NUMA nodes,
+/// which is equal to the number of NUMA nodes that the machine provides. The information about machine's NUMA 
+/// characteristics is provided by the functionality of numa.h lib. In case that numa_regions is different than
+/// HSYNCH_DEFAULT_NUMA_POLICY, the user overides system's default number of NUMA nodes. For example, if  numa_regions = 2
+/// and the machine is equipped with 4 NUMA nodes,then the H-Synch will ignore this and will create a fictitious topology of 
+/// 2 NUMA nodes. This is very usefull in cases of machines that provide many NUMA nodes, but each each of them is equipped
+/// with a small amount of cores. In such a case, the combining degree of H-Synch may be restricted. Thus, creating a 
+/// fictitious topology with restricted number of NUMA nodes gives much better performance. The user usually overides 
+/// HSYNCH_DEFAULT_NUMA_POLICY by setting the '-n' argument in the executable of the benchmarks.
 #define HSYNCH_DEFAULT_NUMA_POLICY 0
 
+/// @brief HalfHSynchNode should not be directly used by the user.
+/// It is internally used for proper alignment of the HSynchNode struct.
 typedef struct HalfHSynchNode {
     struct HalfHSynchNode *next;
     ArgVal arg_ret;
-    int32_t pid;
-    int32_t locked;
-    int32_t completed;
+    uint32_t pid;
+    uint32_t locked;
+    uint32_t completed;
 } HalfHSynchNode;
 
+/// @brief HSynchNode stores the data of an announced request.
 typedef struct HSynchNode {
+    /// @brief Pointer to the next request that has been announced.
     struct HSynchNode *next;
+    /// @brief This variable stores the argument of the request and the return value after the request is applied.
     ArgVal arg_ret;
-    int32_t pid;
-    int32_t locked;
-    int32_t completed;
+    /// @brief The pid of the thread that announced this request.
+    uint32_t pid;
+    /// @brief Whenever it is equal to false, the thread is the combiner; otherwise the thread waits until a combiner apply its request.
+    uint32_t locked;
+    /// @brief If true, the request is applied and the thread returns its return value.
+    uint32_t completed;
+    /// @brief Padding space.
     char align[PAD_CACHE(sizeof(HalfHSynchNode))];
 } HSynchNode;
 
+/// @brief HSynchNodePtr is a struct for padding pointers to nodes of requests.
 typedef union HSynchNodePtr {
     volatile HSynchNode *ptr;
     char pad[CACHE_LINE_SIZE];
 } HSynchNodePtr;
 
+/// @brief HSynchThreadState stores each thread's local state for a single instance of HSynch.
+/// For each instance of HSynch, a discrete instance of HSynchThreadState should be used.
 typedef struct HSynchThreadState {
[…]

[…]
