### 2021-05-08 #3143 @nkallima
Hello @danielskatz and @jarrah42,

In the v2.5.0 branch, you can check the latest updates/improvements on the repository. I will release it to the master branch whenever the desired changes/improvements are in a good shape.

Regarding @jarrah42’s comments,
>  Regarding the Readme. I would like to see a little more detail on how to build the software. E.g. where do you run the make command?

I have added more content in README.md and I have improved some of the content. On how to built the framework you can see more on https://github.com/nkallima/sim-universal-construction/tree/v2.5.0-dev#configuring-compiling-and-installing-the-framework

> Are there other configuration steps?

As a first step, there is no need to configure anything. However, in case someone wants to better explore the performance of the framework, sections https://github.com/nkallima/sim-universal-construction/tree/v2.5.0-dev#configuring-compiling-and-installing-the-framework and https://github.com/nkallima/sim-universal-construction/tree/v2.5.0-dev#performanceoptimizations are a good start.

> Most packages these days provide some kind on installation command that will at least copy the header files and libraries to a specified location. Using CMake or similar will do this for you automatically.

Now, a simple `make install` and `make uninstall` is provided for installing and uninstalling the framework. Please check  https://github.com/nkallima/sim-universal-construction/tree/v2.5.0-dev#configuring-compiling-and-installing-the-framework

> I would definitely like to see complete API documentation, and better documentation on how to use the library.

I have started adding documentation (i.e. manpages) on the API. I have already added manpages for a few of the algorithms, a bit more time is needed in order to have a more complete API documentation.
