### 2021-05-23 462095a
Adding documentation for the API of the Osci combining object, i.e. osci.h.
