### 2021-05-23 fc1c122
Documentation for a simple re-entrant barrier object, i.e. barrier.h.
