### 2021-05-26 444394d
Adding documentation for the API of backoff.h. Documentation improvements on barrier.h.
