### 2021-05-26 756975e
Adding documentation for bench_args.h.
