### 2021-05-30 8eeccc0
Adding documentation for the API of threadtools.h.
