### 2021-05-30 d356048
Adding documentation for the API of stats.h.
