### 2021-06-02 b72af55
Adding documentation for the API of ccqueue.h. Improvements on the documentation of ccsynch.h.
