+++ libconcurrent/includes/ccsynch.h
@@ -1,6 +1,6 @@
 /// @file ccsynch.h
 /// @author Nikolaos D. Kallimanis (nkallima@gmail.com)
-/// @brief This file exposes the API of the CCSynch combining object.
+/// @brief This file exposes the API of the CC-Synch combining object.
 /// An example of use of this API is provided in benchmarks/ccsynchbench.c file.
 ///
 /// For a more detailed description see the original publication:
@@ -39,23 +39,23 @@ typedef struct CCSynchNode {
     char align[PAD_CACHE(sizeof(HalfCCSynchNode))];
 } CCSynchNode;
 
-/// @brief CCSynchThreadState stores each thread's local state for a single instance of CCSynch.
-/// For each instance of CCSynch, a discrete instance of CCSynchThreadState should be used.
+/// @brief CCSynchThreadState stores each thread's local state for a single instance of CC-Synch.
+/// For each instance of CC-Synch, a discrete instance of CCSynchThreadState should be used.
 typedef struct CCSynchThreadState {
     /// @brief pointer to an empty request that would be used for announcing future requests.
     CCSynchNode *next;
-    /// @brief A toggle-bit used by the CCSynch object.
+    /// @brief A toggle-bit used by the CC-Synch object.
     int toggle;
 } CCSynchThreadState;
 
-/// @brief CCSynchStruct stores the state of an instance of the a CCSynch combining object.
+/// @brief CCSynchStruct stores the state of an instance of the a CC-Synch combining object.
 /// CCSynchStruct should be initialized using the CCSynchStructInit function.
 typedef struct CCSynchStruct {
     /// @brief Tail points to the most recently announced request. Initially, it points to NULL.
     volatile CCSynchNode *Tail CACHE_ALIGN;
     /// @brief Pointer to the pool of nodes used by threads in order to announce their requests.
     CCSynchNode *nodes CACHE_ALIGN;
-    /// @brief The number of threads that will use the CCSynch combining object.
+    /// @brief The number of threads that will use the CC-Synch combining object.
     uint32_t nthreads;
 #ifdef DEBUG
     volatile uint64_t counter CACHE_ALIGN;
@@ -64,27 +64,27 @@ typedef struct CCSynchStruct {
 } CCSynchStruct;
 
 
-/// @brief This function initializes an instance of the CCSynch combining object.
+/// @brief This function initializes an instance of the CC-Synch combining object.
 ///
 /// This function should be called once (by a single thread) before any other thread tries to
 /// apply any request by using the CCSynchApplyOp function.
 ///
-/// @param l A pointer to an instance of the CCSynch combining object.
-/// @param nthreads The number of threads that will use the CCSynch combining object.
+/// @param l A pointer to an instance of the CC-Synch combining object.
+/// @param nthreads The number of threads that will use the CC-Synch combining object.
 void CCSynchStructInit(CCSynchStruct *l, uint32_t nthreads);
 
-/// @brief This function should be called once before the thread applies any operation to the CCSynch combining object.
+/// @brief This function should be called once before the thread applies any operation to the CC-Synch combining object.
 ///
-/// @param l A pointer to an instance of the CCSynch combining object.
-/// @param st_thread A pointer to thread's local state of CCSynch.
+/// @param l A pointer to an instance of the CC-Synch combining object.
+/// @param st_thread A pointer to thread's local state of CC-Synch.
 /// @param pid The pid of the calling thread.
 void CCSynchThreadStateInit(CCSynchStruct *l, CCSynchThreadState *st_thread, int pid);
 
 /// @brief This function is called whenever a thread wants to apply an operation to the simulated concurrent object.
 ///
-/// @param l A pointer to an instance of the CCSynch combining object.
-/// @param st_thread A pointer to thread's local state for a specific instance of CCSynch.
[…]
