### 2021-06-02 11ebcb7
Adding documentation for the API of dsmqueue.h.
