### 2021-06-03 b730cdf
Adding documentation for the API of hqueue.h.
