### 2021-06-05 a0f64c5
Adding documentation for the API of ccstack.h. A few documentation fixes.
