+++ libconcurrent/includes/ccqueue.h
@@ -63,7 +63,7 @@ void CCQueueThreadStateInit(CCQueueStruct *object_struct, CCQueueThreadState *lo
 /// @param pid The pid of the calling thread.
 void CCQueueApplyEnqueue(CCQueueStruct *object_struct, CCQueueThreadState *lobject_struct, ArgVal arg, int pid);
 
-/// @brief This function removes (i.e. dequeues) an element from the from of queue and returns its value.
+/// @brief This function removes (i.e. dequeues) an element from the front of the queue and returns its value.
 ///
 /// @param object_struct A pointer to an instance of the CC-Queue concurrent queue implementation.
 /// @param lobject_struct A pointer to thread's local state of CC-Queue.
