+++ libconcurrent/includes/hqueue.h
@@ -67,7 +67,7 @@ void HQueueThreadStateInit(HQueueStruct *object_struct, HQueueThreadState *lobje
 /// @param pid The pid of the calling thread.
 void HQueueApplyEnqueue(HQueueStruct *object_struct, HQueueThreadState *lobject_struct, ArgVal arg, int pid);
 
-/// @brief This function removes (i.e. dequeues) an element from the from of queue and returns its value.
+/// @brief This function removes (i.e. dequeues) an element from the front of the queue and returns its value.
 ///
 /// @param object_struct A pointer to an instance of the H-Queue concurrent queue implementation.
 /// @param lobject_struct A pointer to thread's local state of H-Queue.
