### 2021-06-05 dd4ec59
Adding documentation for the API of dsmstack.h.
