### 2021-06-05 656fadf
Adding documentation for the API of hstack.h.
