### 2021-06-07 4e160dd
Adding documentation for the API of osciqueue.h.
