### 2021-06-07 07a2eec
Adding documentation for the API of oscistack.h. Fixing a few typos in documentation files.
