### 2021-06-14 #3143 @danielskatz
👋 @cedricchevalier19, @mkaguilera, @icalciu, @jamesreinders - Would one of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews of the software and paper here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

As you can see, one of our assigned reviewers was unable to review as he didn't have access to a Linux platform for this, so we now again need to find a second reviewer.
