### 2021-06-15 9806e09
Adding documentation for the API provided by clh.h.
