### 2021-06-15 c47d692
Adding documentation for the API provided by fastrand.h.
