### 2021-06-28 nkallima/sim-universal-construction#31 @williamfgc
[FEATURE] Add user documentation and examples
Not related to a problem per-se, but first time users can benefit for more comprehensive and clear documentation.

**Describe the solution you'd like**
Add user documentation. While there is a substantial amount of documentation in the [README.md](https://github.com/nkallima/sim-universal-construction/blob/master/README.md) a more comprehensive document would benefit the user that want more detail.

GitHub pages, ReadTheDocs, also providing functional examples could help the interested user for reproducibility of the claimed performance and functionality

**Additional context**
Some more specific points to be addressed and be clear about up front in the docs (possibly in README.md) to give a clear scope to the potential user:

- [x] Indicate if this is a library, executable. If it's a C library as it seems to be the case from the API example, please add a prefix_ "namespace" (e.g. `synch_foo`) to library API functions to avoid name clashes
- [x] Indicate this is a Linux-only library
- [x] What's the baseline used for performance claims? Is it measuring strong/weak scalability?
- [x] Illustrate examples comparing this implementation with existing implementations to justify performance (e.g. C++ `std::queue`, `std::unordered_map` for hash-table, C++ <thread> etc?) to users know what to expect in terms of performance.

Related to [JOSS review 3143](https://github.com/openjournals/joss-reviews/issues/3143)
