### 2021-06-28 nkallima/sim-universal-construction#32 @williamfgc
[FEATURE] Add code coverage report in CI
Code coverage is desired to build confidence in the current CI test and to see how much is tested when users want to use a certain feature. It'd also help guidance when adding new tests.

**Describe the solution you'd like**
Either Codecov or Coveralls coverage report + a badge would be a great bonus.

None, but it should be a minor task

**Additional context**
Related to [JOSS review 3143](https://github.com/openjournals/joss-reviews/issues/3143)
