### 2021-06-28 #3143 @williamfgc
@nkallima I just opened issues above requesting some enhancements based on my initial review. Overall, great job and thanks for putting together a piece of functional software out of your research. I still need to go through some of the software functionalities, but these are initial requests/comments/feedback mostly to make the software more user-friendly. Bonus: *great CI and tests*.
