### 2021-06-29 #3143 @nkallima
@williamfgc Thank you for your constructive comments and for the time devoted to review the repository. I am going to address soon the open issues and your comments directly in the repository.

> @nkallima I just opened issues above requesting some enhancements based on my initial review. Overall, great job and thanks for putting together a piece of functional software out of your research. I still need to go through some of the software functionalities, but these are initial requests/comments/feedback mostly to make the software more user-friendly. Bonus: _great CI and tests_.
