### 2021-07-02 nkallima/sim-universal-construction#32 @nkallima
Thanks for your feedback!
Yes, I think it is a good idea to have a coverage report for the validation tests. I will check both Codecov and Coveralls in order to see which better suits the project.

BTW, in the v2.5.0 branch, I push all the improvements for the JOSS review. I will release it to the master branch whenever the desired changes/improvements are in a good shape.
