### 2021-07-02 nkallima/sim-universal-construction#31 @nkallima
>Describe the solution you'd like
>Add user documentation. While there is a substantial amount of documentation in the >README.md a more comprehensive document would benefit the user that want more detail.

In the v2.5.0 branch (it will be released to the master branch whenever it is in a good shape), I have started creating documentation for the provided API (i.e. all the header files under the libconcurrent/includes/ directory). Until now, a substantial set of the provided API is documented; however, a few files remain to be documented.

>Describe alternatives you've considered
>GitHub pages, ReadTheDocs, also providing functional examples could help the interested user for reproducibility of the claimed performance and functionality

Currently, the provided documentation is produced by Doxygen creating man-pages  (check https://github.com/nkallima/sim-universal-construction/tree/v2.5.0-dev#configuring-compiling-and-installing-the-framework for more), which is a bit old-fashioned.
I will check if there is an easy way to move a part of the documentation to GitHub pages or ReadTheDocs as you suggested.

> Indicate if this is a library, executable. If it's a C library as it seems to be the case from the API example, please add a prefix_ "namespace" (e.g. synch_foo) to library API functions to avoid name clashes

This is a good idea, at least for the basic functionality provided by the runtime and the primitives. However, I’m a bit skeptical to adopt it for all the provided concurrent algorithms, since the API of some data-structures is a bit old and I hesitate to change the naming scheme. In any case, it’s a good idea to start changing the runtime and primitives naming.
