### 2021-07-07 #3143 @nkallima
> wave @nkallima - How are these updates going?

@danielskatz I have started working on the changes that williamfgc has proposed. Also, I have replied to a few comments directly in the repository's issues.
