### 2021-07-16 3692cdd
Change a bit codecov limits in order to be more realistic.
