### 2021-07-16 f4bf8dd
Change a bit codecov limits in order to be more realistic.
