### 2021-07-16 02d41f4
Adding configuration for codecov.
