+++ .codecov.yml
@@ -1,7 +1,7 @@
 coverage:
   precision: 2
   round: down
-  range: "60...100"
+  range: "50...100"
 
   status:
     project: no
