+++ README.md
@@ -1,4 +1,4 @@
-[![check-build](https://github.com/nkallima/sim-universal-construction/actions/workflows/check-build.yml/badge.svg)](https://github.com/nkallima/sim-universal-construction/actions/workflows/check-build.yml) [![validate](https://github.com/nkallima/sim-universal-construction/actions/workflows/validate.yml/badge.svg)](https://github.com/nkallima/sim-universal-construction/actions/workflows/validate.yml)
+[![check-build](https://github.com/nkallima/sim-universal-construction/actions/workflows/check-build.yml/badge.svg)](https://github.com/nkallima/sim-universal-construction/actions/workflows/check-build.yml) [![validate](https://github.com/nkallima/sim-universal-construction/actions/workflows/validate.yml/badge.svg)](https://github.com/nkallima/sim-universal-construction/actions/workflows/validate.yml) [![codecov](https://codecov.io/gh/nkallima/sim-universal-construction/branch/codecov/graph/badge.svg?token=1V8A6BOABM)](https://codecov.io/gh/nkallima/sim-universal-construction)
 
 <p align="center">
     <img src="resources/logo_synch.png" alt="The Synch Framework" width="80%">
