+++ Makefile
@@ -10,6 +10,9 @@ all:
 debug:
 	make $(ARCH) D_ARGS="-DDEBUG"
 
+codecov:
+	make $(ARCH) D_ARGS="-DDEBUG -ftest-coverage -fprofile-arcs --coverage -O0"
+
 x86_64:
 	make -f Makefile.generic -j $(NPROC) COMPILER=$(CC) LDLIBS='$(LIBS)' CFLAGS='-Wall -std=gnu89 -Ofast -fPIC -flto -ftree-vectorize -finline-functions -march=native -mtune=native'
 
