+++ .codecov.yml
@@ -1,0 +1,20 @@
+coverage:
+  precision: 2
+  round: down
+  range: "50...100"
+
+  status:
+    project: no
+    patch: yes
+    changes: no
+
+comment:
+  layout: "header, diff, changes, tree"
+  behavior: default
+
+ignore:
+  - "resources/*"
+  - "*.sh"
+  - "*.md"
+  - "Makefile*"
+  - "build/*"
