### 2021-07-16 nkallima/sim-universal-construction#32 @nkallima
@williamfgc

I added a coverage report for validation checks to v2.5.0-dev branch (please check https://github.com/nkallima/sim-universal-construction/tree/v2.5.0-dev). It seems that the code-coverage is about 66% at this point.
