### 2021-07-16 88d8c32
Update deploy_docs.sh
