+++ .travis.yml
@@ -1,8 +1,8 @@
 sudo: false
 
 branches:
-  except:
-    - gh-pages
+  only:
+    - v2.5.0-dev
 
 addons:
   apt:
