### 2021-07-19 cdcce83
Improvements on the deploy script.
