+++ .travis.yml
@@ -1,17 +1,0 @@
-sudo: false
-
-branches:
-  only:
-    - v2.5.0-dev
-
-addons:
-  apt:
-    packages:
-      - doxygen
-      - doxygen-doc
-      - doxygen-latex
-      - doxygen-gui
-      - graphviz
-
-script:
-  - ./deploy_docs.sh
