### 2021-07-19 #3143 @danielskatz
👋 All, I was on vacation last week, and am now trying to catch up.
@jarrah42 & @williamfgc - can you let me know what your reviews are blocked on at this point?
@nkallima - can you let me know where you are in responding to reviewer issues?
