### 2021-07-19 06a5575
Adding a script for deploying documentation in github-pages.
