### 2021-07-19 b362d11
Improvement on the deploy documentation script.
