### 2021-07-19 #3143 @nkallima
@danielskatz, @jarrah42 & @williamfgc
Now most of the API documentation is provided under the github-pages (https://nkallima.github.io/sim-universal-construction/index.html).
Although the most important parts are documented, a small part of the API remains to be documented.
