### 2021-07-20 61a4000
Adding documentation for the Oyama's combining object.
