+++ .codecov.yml
@@ -5,7 +5,7 @@ coverage:
 
   status:
     project: no
-    patch: yes
+    patch: no
     changes: no
 
 comment:
