### 2021-07-20 b8c109b
Update resources/deploy_docs.sh.
