### 2021-07-20 3f2b286
Adding documentation for the LCRQ concurrent queue implementation.
