+++ libconcurrent/includes/lcrq.h
@@ -1,15 +1,18 @@
-// This is a slightly modified version of the LCRQ queue implementation by Adam Morrison
-// and Yehuda Afek provided in http://mcg.cs.tau.ac.il/projects/lcrq.
-// LCRQ is presented in "Fast concurrent queues for x86 processors", by Adam Morrison,
-// and Yehuda Afek, in PPoPP 2013.
-// This modified version has some minor modifications in order to be compatible with the 
-// latest version of the Synch framework.
-// Notice that this algorithm performs exceptionally well in X86 machines that natively
-// support 128-bit Compare&Swap instructions. It is also tested in ARM-V8 and in RISCV machines.
-// This modified version seems to have very similar performance as the original one.
-// The original version of LCRQ is distributed under the New BSD Licence (LGPL-2.1 is 
-// compatible with New BSD Licence, more on https://www.gnu.org/licenses/license-compatibility.html).
-
+/// @file lcrq.h
+/// @author Adam Morrison and Yehuda Afek
+/// @brief This is a slightly modified version of the LCRQ queue implementation by Adam Morrison
+/// and Yehuda Afek provided in http://mcg.cs.tau.ac.il/projects/lcrq.
+///
+/// LCRQ is presented in "Fast concurrent queues for x86 processors", by Adam Morrison,
+/// and Yehuda Afek, in PPoPP 2013.
+/// This modified version has some minor modifications in order to be compatible with the 
+/// latest version of the Synch framework.
+/// Notice that this algorithm performs exceptionally well in X86 machines that natively
+/// support 128-bit Compare&Swap instructions. It is also tested in ARM-V8 and in RISCV machines.
+/// This modified version seems to have very similar performance as the original one.
+/// The original version of LCRQ is distributed under the New BSD Licence (LGPL-2.1 is 
+/// compatible with New BSD Licence, more on https://www.gnu.org/licenses/license-compatibility.html).
+/// @copyright Copyright (c) 2021
 
 // Copyright (c) 2013, Adam Morrison and Yehuda Afek.
 // All rights reserved.
@@ -74,8 +77,12 @@ typedef struct RingQueue {
     RingNode array[RING_SIZE] S_CACHE_ALIGN;
 } RingQueue;
 
+/// @brief LCRQStruct stores the state of an instance of the LCRQ concurrent queue implementation.
+/// LCRQStruct should be initialized using the LCRQStructInit function.
 typedef struct LCRQStruct {
+    /// @brief A pointer to the head Ring.
     RingQueue *head;
+    /// @brief A pointer to the tail Ring.
     RingQueue *tail;
 #ifdef DEBUG
     uint64_t closes;
@@ -83,6 +90,8 @@ typedef struct LCRQStruct {
 #endif
 } LCRQStruct;
 
+/// @brief LCRQThreadState stores each thread's local state for a single instance of LCRQ.
+/// For each instance of LCRQ, a discrete instance of LCRQThreadState should be used.
 typedef struct LCRQThreadState {
     RingQueue *nrq;
 #ifdef HAVE_HPTRS
@@ -94,7 +103,34 @@ typedef struct LCRQThreadState {
 #endif
 } LCRQThreadState;
 
+
+/// @brief This function initializes an instance of the LCRQ concurrent queue implementation.
+/// This function should be called once (by a single thread) before any other thread tries to
+/// apply any enqueue or dequeue operation.
+///
+/// @param queue A pointer to an instance of the LCRQ concurrent queue implementation.
+/// @param nthreads The number of threads that will use the LCRQ concurrent queue implementation.
 void LCRQInit(LCRQStruct *queue, uint32_t nthreads);
+
[…]
