### 2021-07-21 cfa66c4
Adding documentation for the lock-free unversal construction (i.e. lfuobject.h).
