### 2021-07-21 cb75b11
Adding documentation for the fam.h (i.e. a serial implementation of a Fetch&Multiply objec).
