### 2021-07-21 cc2e6e2
Expose documentation of config.h to doxygen.
