+++ libconcurrent/config.h
@@ -1,88 +1,73 @@
+/// @file config.h
+/// @author Nikolaos D. Kallimanis (nkallima@gmail.com)
+/// @brief This file provides important parameters and consstant for the benchmarks and the  library.
 #ifndef _CONFIG_H_
 #define _CONFIG_H_
 
 #include <stdint.h>
 
-// Definition: MAX_WORK
-// --------------------
-// Define the default value of maximum local work that each thread executes between two consecutive of the
-// benchmarked operation. A zero value means there is no work between two consecutive calls.
-// Large values usually reduce system's contention, i.e. threads perform operations less frequentely.
-// In constrast, small values (but not zero) increase system's contention. Please avoid to set this value
-// equal to zero, since some algorithms may produce unreallistically high performan (i.e. long runs
-// and unrealistic low numbers of cache misses).
-// Default value is 64.
+/// @brief Define the default value of maximum local work that each thread executes between two consecutive of the
+/// benchmarked operation. A zero value means there is no work between two consecutive calls.
+/// Large values usually reduce system's contention, i.e. threads perform operations less frequentely.
+/// In constrast, small values (but not zero) increase system's contention. Please avoid to set this value
+/// equal to zero, since some algorithms may produce unreallistically high performan (i.e. long runs
+/// and unrealistic low numbers of cache misses).
+/// Default value is 64.
 #ifndef MAX_WORK
 #    define MAX_WORK               64
 #endif
 
-// definition: RUNS
-// ----------------
-// Define the default total number of the executed operations. Notice that benchmarks for stacks and queues
-// execute RUNS pairs of operations (i.e. pairs of push/pops or pairs of enqueues/dequeues).
-// Default value is 1000000.
+/// @brief Define the default total number of the executed operations. Notice that benchmarks for stacks and queues
+/// execute RUNS pairs of operations (i.e. pairs of push/pops or pairs of enqueues/dequeues).
+/// Default value is 1000000.
 #define RUNS                       1000000
 
-// Definition: DEBUG
-// -----------------
-// Define DEBUG,  in case you want to debug some parts of the code or to get some  useful performance statistics.
-// Note that the validation.sh script enables this definition by default. In some cases, this may introduces
-// some performance loses. Thus, in case you want to perform benchmarking keeps this undefined.
-// By default, this flag is disabled.
+/// @brief Define DEBUG,  in case you want to debug some parts of the code or to get some  useful performance statistics.
+/// Note that the validation.sh script enables this definition by default. In some cases, this may introduces
+/// some performance loses. Thus, in case you want to perform benchmarking keeps this undefined.
+/// By default, this flag is disabled.
 //#define DEBUG
 
-// Definition: DISABLE_BACKOFF
-// ---------------------------
-// This definition disbales backoff in all algorithms that are using it for reducing system's contention.
-// By default, this flag is disabled.
+/// @brief This definition disbales backoff in all algorithms that are using it for reducing system's contention.
+/// By default, this flag is disabled.
 //#define DISABLE_BACKOFF
 
 #define Object                     int64_t
 
-// Definition: RetVal
-// ------------------
-// Define the type of the return value that most operations in the provided benchmarks return.
-// For instance, the return values of all pop and dequeue operations (implemented by the provided stacks and queue
-// objects) return a value of type RetVal. It is assumed that the target architecture is able to atomically read/write 
-// this type If this type is of 32 or 64 bits could be atomically read or written by most machine architectures.
-// However, a value of 128 bits or more may not be supported (in most cases x86_64 supports types of 128 bits).
-// By default, this definition is equal to int64_t
+/// @brief Define the type of the return value that most operations in the provided benchmarks return.
[…]
