### 2021-07-21 4e109b7
Adding documentation for the Sim universal construction.
