+++ libconcurrent/includes/sim.h
@@ -1,3 +1,12 @@
+/// @file sim.h
+/// @author Nikolaos D. Kallimanis (nkallima@gmail.com)
+/// @brief This file exposes the API of the P-Sim (or Sim) universal construction.
+/// An example of use of this API is provided in benchmarks/simbench.c file.
+///
+/// For a more detailed description see the original publication:
+/// Panagiota Fatourou, and Nikolaos D. Kallimanis. "A highly-efficient wait-free universal construction". 
+/// Proceedings of the twenty-third annual ACM symposium on Parallelism in algorithms and architectures (SPAA), 2011.
+/// @copyright Copyright (c) 2021
 #ifndef _SIM_H_
 #define _SIM_H_
 
@@ -7,12 +16,16 @@
 #include <tvec.h>
 #include <fam.h>
 
+/// @brief This constant controls the size of pool of SimObjectState structs that each thread maintains.
+/// This should be a small integer (i.e, 2 or 4) in order to avoid excess memory consumption.
 #define _SIM_LOCAL_POOL_SIZE_ 4
 
 #if _SIM_LOCAL_POOL_SIZE_ < 2
 #    error SIM universal construction is improperly configured
 #endif
 
+/// @brief This should not used directely by user. 
+/// It is used for padding the SimObjectState struct appropriately.
 typedef struct HalfSimObjectState {
     RetVal *ret;
     ToggleVector applied;
@@ -24,9 +37,13 @@ typedef struct HalfSimObjectState {
     uint64_t __flex[1];
 } HalfSimObjectState;
 
+/// @brief This struct stores the data for a copy of the simulated object's state.
 typedef struct SimObjectState {
+    /// @brief A pointer to the array of return values.
     RetVal *ret;
+    /// @brief The applied vector of toggles.
     ToggleVector applied;
+    /// @brief The actual data of the simulated object's state.
     ObjectState state CACHE_ALIGN;
 #ifdef DEBUG
     int counter;
@@ -36,8 +53,12 @@ typedef struct SimObjectState {
     char pad[PAD_CACHE(sizeof(HalfSimObjectState))];
 } SimObjectState;
 
+/// @brief A macro for calculating the size of the SimObjectState struct for a specific amount of threads.
 #define SimObjectStateSize(nthreads) (sizeof(SimObjectState) + _TVEC_VECTOR_SIZE(nthreads) + (nthreads) * sizeof(RetVal))
 
+/// @brief pointer_t should not used directely by user. This struct is used by Sim for pointing to the 
+/// most rescent and valid copy of the simulated object's state. It also contains a 40-bit sequence number
+/// for avoiding the ABA problem.
 typedef union pointer_t {
     struct StructData {
         uint64_t seq : 40;
@@ -46,31 +67,66 @@ typedef union pointer_t {
     volatile uint64_t raw_data;
 } pointer_t;
 
+/// @brief SimThreadState stores each thread's local state for a single instance of Sim.
+/// For each instance of Sim, a discrete instance of SimThreadState should be used.
 typedef struct SimThreadState {
     ToggleVector mask;
     ToggleVector toggle;
     ToggleVector my_bit;
     ToggleVector diffs;
     ToggleVector l_toggles;
+    /// @brief The next available free copy of object's state.
     int local_index;
+    /// @brief Current backoff value.
     int backoff;
 } SimThreadState;
 
+/// @brief SimStruct stores the state of an instance of the a Sim combining object.
+/// SimStruct should be initialized using the SimStructInit function.
 typedef struct SimStruct {
+    /// @brief Pointer to a SimObjectState structs that contains the most recent and valid copy of simulated object's state.
     volatile pointer_t sp;
 
[…]
