### 2021-07-21 ef8a39c
Adding documentations for the SimQueue cpncurrent queue implementation.
