+++ benchmarks/simqueuebench.c
@@ -49,7 +49,7 @@ static void *Execute(void *Arg) {
 int main(int argc, char *argv[]) {
     parseArguments(&bench_args, argc, argv);
     queue = getAlignedMemory(CACHE_LINE_SIZE, sizeof(SimQueueStruct));
-    SimQueueInit(queue, bench_args.nthreads, bench_args.backoff_high);
+    SimQueueStructInit(queue, bench_args.nthreads, bench_args.backoff_high);
 
     BarrierSet(&bar, bench_args.nthreads);
     StartThreadsN(bench_args.nthreads, Execute, bench_args.fibers_per_thread);
