+++ libconcurrent/includes/simqueue.h
@@ -1,3 +1,12 @@
+/// @file simqueue.h
+/// @author Nikolaos D. Kallimanis (nkallima@gmail.com)
+/// @brief This file exposes the API of the SimQueue universal construction.
+/// An example of use of this API is provided in benchmarks/simqueuebench.c file.
+///
+/// For a more detailed description see the original publication:
+/// Panagiota Fatourou, and Nikolaos D. Kallimanis. "A highly-efficient wait-free universal construction". 
+/// Proceedings of the twenty-third annual ACM symposium on Parallelism in algorithms and architectures (SPAA), 2011.
+/// @copyright Copyright (c) 2021
 #ifndef _SIMQUEUE_H_
 #define _SIMQUEUE_H_
 
@@ -7,11 +16,17 @@
 #include <sim.h>
 #include <pool.h>
 
+/// @brief This struct stores the data for the state of the enqueue instance. Used by the threads that perform enqueue operations.
 typedef struct EnqState {
+    /// @brief The applied vector of toggles.
     ToggleVector applied;
+    /// @brief In case you want to copy this struct, start from here.
     uint64_t copy_point;
+    /// @brief A pointer to the tail of the queue.
     Node *tail CACHE_ALIGN;
+    /// @brief A pointer to the first element of the last batch of applied enqueue operations.
     Node *first;
+    /// @brief A pointer to the last element of the last batch of applied enqueue operations.
     Node *last;
 #ifdef DEBUG
     int32_t counter;
@@ -19,12 +34,18 @@ typedef struct EnqState {
     uint64_t __flex[1];
 } EnqState;
 
+/// @brief A macro for calculating the size of the EnqState struct for a specific amount of threads.
 #define EnqStateSize(N) (sizeof(EnqState) + _TVEC_VECTOR_SIZE(N))
 
+/// @brief This struct stores the data for the state of the dequeue instance. Used by the threads that perform dequeue operations.
 typedef struct DeqState {
+    /// @brief The applied vector of toggles.
     ToggleVector applied;
+    /// @brief In case you want to copy this struct, start from here.
     uint64_t copy_point;
+    /// @brief A pointer to the head of the queue.
     Node *head;
+    /// @brief A pointer to the array of return values.
     RetVal *ret;
 #ifdef DEBUG
     int32_t counter;
@@ -32,9 +53,11 @@ typedef struct DeqState {
     uint64_t __flex[1];
 } DeqState;
 
+/// @brief A macro for calculating the size of the DeqState struct for a specific amount of threads.
 #define DeqStateSize(N) (sizeof(DeqState) + _TVEC_VECTOR_SIZE(N) + (N) * sizeof(RetVal))
 
-// Each thread owns a private copy of the following variables
+/// @brief SimQueueThreadState stores each thread's local state for a single instance of SimQueue.
+/// For each instance of SimQueue, a discrete instance of SimQueueThreadState should be used.
 typedef struct SimQueueThreadState {
     ToggleVector mask;
     ToggleVector deq_toggle;
@@ -43,30 +66,72 @@ typedef struct SimQueueThreadState {
     ToggleVector my_enq_bit;
     ToggleVector diffs;
     ToggleVector l_toggles;
+    /// @brief A pool of Node structs used for fast allocation during the enqueue operations.
     PoolStruct pool_node;
+    /// @brief The next available free copy of EnqState that could be used on an enqueue operation.
     int deq_local_index;
+    /// @brief The next available free copy of DeqState that could be used on a dequeue operation.
     int enq_local_index;
-    int backoff;
+    /// @brief The maximum backoff value.
+    int max_backoff;
 } SimQueueThreadState;
 
+/// @brief SimQueueStruct stores the state of an instance of the SimQueue.
+/// SimQueueStruct should be initialized using the SimQueueStructInit function.
[…]
