### 2021-07-22 06ba370
Adding documentation for the SimStack concurrent stack implementation. Improving documentation of simqueue.h.
