+++ libconcurrent/includes/simstack.h
@@ -1,3 +1,12 @@
+/// @file simstack.h
+/// @author Nikolaos D. Kallimanis (nkallima@gmail.com)
+/// @brief This file exposes the API of the SimStack concurrent stack implementation.
+/// An example of use of this API is provided in benchmarks/simstackbench.c file.
+///
+/// For a more detailed description see the original publication:
+/// Panagiota Fatourou, and Nikolaos D. Kallimanis. "A highly-efficient wait-free universal construction". 
+/// Proceedings of the twenty-third annual ACM symposium on Parallelism in algorithms and architectures (SPAA), 2011.
+/// @copyright Copyright (c) 2021
 #ifndef _SIMSTACK_H_
 #define _SIMSTACK_H_
 
@@ -7,6 +16,7 @@
 #include <pool.h>
 #include <queue-stack.h>
 
+/// @brief This struct is  used for padding the SimStackState struct appropriately.
 typedef struct HalfSimStackState {
     ToggleVector applied;
     Object *ret;
@@ -17,20 +27,29 @@ typedef struct HalfSimStackState {
     uint64_t __flex[1];
 } HalfSimStackState;
 
+/// @brief This struct stores the data about the state of a SimStack instance.
 typedef struct SimStackState {
+    /// @brief The applied vector of toggles.
     ToggleVector applied;
+    /// @brief A pointer to the array of return values.
     Object *ret;
+    /// @brief A pointer to the head node of the stack.
     Node *head;
 #ifdef DEBUG
     int counter;
 #endif
     uint64_t __flex[1];
+    /// @brief Padding space.
     char pad[PAD_CACHE(sizeof(HalfSimStackState))];
 } SimStackState;
 
+/// @brief A macro for calculating the size of the SimStackState struct for a specific amount of threads.
 #define SimStackStateSize(nthreads) (sizeof(SimStackState) + _TVEC_VECTOR_SIZE(nthreads) + nthreads * sizeof(Object))
 
+/// @brief SimStackThreadState stores each thread's local state for a single instance of SimStack.
+/// For each instance of SimStack, a discrete instance of SimStackThreadState should be used.
 typedef struct SimStackThreadState {
+    /// @brief A pool of Node structs used for fast allocation during push operations.
     PoolStruct pool;
     ToggleVector mask;
     ToggleVector toggle;
@@ -38,23 +57,64 @@ typedef struct SimStackThreadState {
     ToggleVector diffs;
     ToggleVector l_toggles;
     ToggleVector pops;
+    /// @brief The next available free copy of SimStackState that could be used while applying push and/or pop operations.
     int local_index;
+    /// @brief The maximum backoff value.
     int backoff;
 } SimStackThreadState;
 
+/// @brief SimStackStruct stores the state of an instance of the SimStack.
+/// SimStackStruct should be initialized using the SimStackStructInit function.
 typedef struct SimStackStruct {
+    /// @brief Pointer to an array, where threads announce only the requests that want to perform to the object.
     ArgVal *announce;
+    /// @brief An array of pools (one pool per thread) of SimStackState structs.
     SimStackState **pool;
+    /// @brief The number of threads that use this instance of SimStack.
     uint32_t nthreads;
+    /// @brief The maximum backoff value.
     int MAX_BACK;
+    /// @brief A pointer to the head node of the stack.
     volatile Node *head CACHE_ALIGN;
+    /// @brief Pointer to a SimStackState struct that contains the most recent and valid copy of this SimStack instance.
     volatile pointer_t sp CACHE_ALIGN;
+    /// @brief A vector of toggle bits used to detect which announced operations are applied or not.
[…]
