### 2021-07-23 8c03346
Adding documentation for pool.h.
