### 2021-07-23 5c9eb33
Fixing a few inconsistencies in the documentation.
