+++ libconcurrent/includes/lfuobject.h
@@ -40,14 +40,12 @@ typedef struct LFUObjectThreadState {
 /// apply any request by using the LFUObjectApplyOp function.
 ///
 /// @param l A pointer to an instance of the lock-free universal construction.
-/// @param nthreads The number of threads that will use the lock-free universal construction.
+/// @param value The initial value that the simulated object should have.
 void LFUObjectInit(LFUObjectStruct *l, ArgVal value);
 
 /// @brief This function should be called once before the thread applies any operation to the lock-free universal construction.
 ///
-/// @param l A pointer to an instance of the lock-free universal construction.
 /// @param th_state A pointer to thread's local state of lock-free.
-/// @param pid The pid of the calling thread.
 /// @param min_back The minimum value for backoff (in most cases 0 is a good start).
 /// @param max_back The maximum value for backoff (usually this is much lower than 100).
 void LFUObjectThreadStateInit(LFUObjectThreadState *th_state, int min_back, int max_back);
@@ -57,7 +55,6 @@ void LFUObjectThreadStateInit(LFUObjectThreadState *th_state, int min_back, int 
 /// @param l A pointer to an instance of the lock-free universal construction.
 /// @param th_state A pointer to thread's local state for a specific instance of lock-free.
 /// @param sfunc A serial function that the lock-free instance should execute, while applying requests announced by active threads.
-/// @param state A pointer to the state of the simulated object.
 /// @param arg The argument of the request that the thread wants to apply.
 /// @param pid The pid of the calling thread.
 /// @return RetVal The return value of the applied request.
