+++ libconcurrent/includes/clhhash.h
@@ -1,3 +1,17 @@
+/// @file clhhash.h
+/// @brief This file exposes the API of the CLH-Hash concurrent hash-table implementation based on the CLH queue locks [1, 2].
+///
+/// This hash-table implementation uses an array of a fixed amount of cells that it is defined during the initialization of the data-structure. 
+/// Each cell contains at most one <key,value> pair and a pointer to a next cell. For handling collisions this hash-table uses separate chaining.
+/// Each cell of the array is protected by a single CLH lock. Whenever a thread wants to search, insert or delete a <key,value> pair in a
+/// specific cell, it should first acquire the lock of the cell. Since this hash-table implementation is not expandable 
+/// (the initial size of the array is fixed), it may become very slow in cases of storing big amounts of <key,value> pairs.
+/// An example of use of this API is provided in benchmarks/clhhashbench.c file.
+///
+/// References
+/// ----------
+/// [1]. Travis S. Craig. "Building FIFO and priority-queueing spin locks from atomic swap". Technical Report TR 93-02-02, Department of Computer Science, University of Washington, February 1993.
+/// [2]. Peter Magnusson, Anders Landin, and Erik Hagersten. "Queue locks on cache coherent multiprocessors". Parallel Processing Symposium, 1994. Proceedings., Eighth International. IEEE, 1994.
 #ifndef _CLHHASH_H_
 #define _CLHHASH_H_
 
@@ -7,35 +21,97 @@
 #include <pool.h>
 #include <clh.h>
 
+/// @brief A node (i.e. cell) for the linked-list of cells of <key,value> pair.
+/// This should not directely accessed-used by the user.
 typedef struct HashNode {
+    /// @brief Pointer to next node (i.e. cell) of the linked list.
     struct HashNode *next;
+    /// @brief The key of the <key,value> pair for the specific node.
     int64_t key;
+    /// @brief The value of the <key,value> pair for the specific node.
     int64_t value;
 } HashNode;
 
+/// @brief HashOperations describes a hash-table operation, i.e. search, insert or delete.
+/// This should not directely accessed-used by the user.
 typedef struct HashOperations {
+    /// @brief The key of the <key,value> pair for the specific request.
     int64_t key;
+    /// @brief The value of the <key,value> pair for the specific request.
     int64_t value;
+    /// @brief A new cell for the insert operations.
     HashNode *node;
-    int32_t bucket;
+    /// @brief This field stores result of the hash function.
+    int32_t cell;
+    /// @brief The type of hash-table operation, i.e. insert, delete or search.
     int32_t op;
 } HashOperations;
 
+/// @brief CLHHash stores the state of an instance of the a CLH-Hash concurrent hash-table.
+/// CLHHash should be initialized using the CLHHashStructInit function.
 typedef struct CLHHash {
+    /// @brief A pointer to the announce array.
     HashOperations *announce;
+    /// @brief A pointer to an array of CLH locks.
     CLHLockStruct **synch CACHE_ALIGN;
-    ptr_aligned_t *buckets;
+    /// @brief A pointer to an array of the cells.
+    ptr_aligned_t *cells;
+    /// @brief The size in terms of cells of the hash-table.
     int size;
 } CLHHash;
 
+/// @brief CLHHashThreadState stores each thread's local state for a single instance of CLH-Hash.
+/// For each instance of CLH-Hash, a discrete instance of CLHHashThreadState should be used.
 typedef struct CLHHashThreadState {
+    /// @brief A pool of nodes for fast memory allocation.
     PoolStruct pool;
 } CLHHashThreadState;
 
-inline void CLHHashInit(CLHHash *hash, int hash_size, int nthreads);
[…]
