+++ libconcurrent/includes/dsmhash.h
@@ -1,3 +1,16 @@
+/// @file dsmhash.h
+/// @brief This file exposes the API of the DSM-Hash concurrent hash-table implementation based on the DSM-Synch combining technique [1].
+///
+/// This hash-table implementation uses an array of a fixed amount of cells that it is defined during the initialization of the data-structure. 
+/// Each cell contains at most one <key,value> pair and a pointer to a next cell. For handling collisions this hash-table uses separate chaining.
+/// For each cell there is a single instance of DSM-Synch. Whenever a thread wants to search, insert or delete a <key,value> pair in a
+/// specific cell, it should use the corresponding DSM-Synch instance. Since this hash-table implementation is not expandable 
+/// (the initial size of the array is fixed), it may become very slow in cases of storing big amounts of <key,value> pairs.
+/// An example of use of this API is provided in benchmarks/dsmhashbench.c file.
+///
+/// References
+/// ----------
+/// [1]. Panagiota Fatourou, and Nikolaos D. Kallimanis. "Revisiting the combining synchronization technique". ACM SIGPLAN Notices. Vol. 47. No. 8. ACM, PPoPP 2012.
 #ifndef _DSMHASH_H_
 #define _DSMHASH_H_
 
@@ -7,36 +20,99 @@
 #include <pool.h>
 #include <types.h>
 
+/// @brief A node (i.e. cell) for the linked-list of cells of <key,value> pair.
+/// This should not directely accessed-used by the user.
 typedef struct HashNode {
+    /// @brief Pointer to next node (i.e. cell) of the linked list.
     struct HashNode *next;
+    /// @brief The key of the <key,value> pair for the specific node.
     int64_t key;
+    /// @brief The value of the <key,value> pair for the specific node.
     int64_t value;
 } HashNode;
 
+/// @brief HashOperations describes a hash-table operation, i.e. search, insert or delete.
+/// This should not directely accessed-used by the user.
 typedef struct HashOperations {
+    /// @brief The key of the <key,value> pair for the specific request.
     int64_t key;
+    /// @brief The value of the <key,value> pair for the specific request.
     int64_t value;
+    /// @brief A new cell for the insert operations.
     HashNode *node;
-    int32_t bucket;
+    /// @brief This field stores result of the hash function.
+    int32_t cell;
+    /// @brief The type of hash-table operation, i.e. insert, delete or search.
     int32_t op;
 } HashOperations;
 
+/// @brief DSMHash stores the state of an instance of the a DSM-Hash concurrent hash-table.
+/// DSMHash should be initialized using the DSMHashStructInit function.
 typedef struct DSMHash {
+    /// @brief A pointer to the announce array.
     HashOperations *announce;
+    /// @brief A pointer to an array of DSM locks.
     DSMSynchStruct *synch CACHE_ALIGN;
-    ptr_aligned_t *buckets;
+    /// @brief A pointer to an array of the cells.
+    ptr_aligned_t *cells;
+    /// @brief The size in terms of cells of the hash-table.
     int size;
 } DSMHash;
 
+/// @brief DSMHashThreadState stores each thread's local state for a single instance of DSM-Hash.
+/// For each instance of DSM-Hash, a discrete instance of DSMHashThreadState should be used.
 typedef struct DSMHashThreadState {
+    /// @brief A pointer to the thread's local state for the DSM-Synch instance.
     DSMSynchThreadState *th_state;
+    /// @brief A pool of nodes for fast memory allocation.s
     PoolStruct pool;
 } DSMHashThreadState;
 
-inline void DSMHashInit(DSMHash *hash, int hash_size, int nthreads);
-inline void DSMHashThreadStateInit(DSMHash *hash, DSMHashThreadState *th_state, int hash_size, int pid);
[…]

[…]
