### 2021-07-26 nkallima/sim-universal-construction#34 @williamfgc
[FEATURE] Improve Contributing documentation
Contributing is not clear for external engagement.

**Describe the solution you'd like**
Clarify the desired workflow (fork-branch ? Should `master` (now `main` be up-to-date?).
It's not clear if someone would like to do a Pull Request where things should go.

Just expand Contributing.md

**Additional context**
Part of [JOSS review issue 3143](https://github.com/openjournals/joss-reviews/issues/3143)
