### 2021-07-26 nkallima/sim-universal-construction#35 @williamfgc
[FEATURE] Switch master to main for default branch
GitHub has demoted the use of the `master` branch in favor of `main` going forward.

**Describe the solution you'd like**
It should be an easy fix: https://github.com/github/renaming

None

**Additional context**
[GitHub docs for changing default branch](https://docs.github.com/en/github/administering-a-repository/managing-branches-in-your-repository/changing-the-default-branch)
Part of [JOSS review issue 3143](https://github.com/openjournals/joss-reviews/issues/3143)
