### 2021-07-26 nkallima/sim-universal-construction#32 @williamfgc
@nkallima thanks, nice job. Just as a suggestion, automatic coverage report can be a badge in your ReadMe.md
