### 2021-07-26 nkallima/sim-universal-construction#31 @williamfgc
> However, I’m a bit skeptical to adopt it for all the provided concurrent algorithms

Fair point, let me clarify my initial request. The public facing API should have a prefix (what users are supposed to "consume"), everything that has a private scope is fine. Let me know if this makes any sense. Thanks for addressing the other points!
