### 2021-07-26 nkallima/sim-universal-construction#32 @nkallima
@williamfgc thanks for your comments and your reply. The badge is already there, in the README.md of v2.5.0-dev branch. In the v2.5.0 branch, I'm pushing all the improvements for the JOSS review. I will merge to the main branch whenever the desired changes/improvements are in a good shape.
