### 2021-07-26 nkallima/sim-universal-construction#32 @nkallima
@williamfgc, I replied for #34 in its own issue. Is it OK for you to close the current issue?
