### 2021-07-26 nkallima/sim-universal-construction#32 @nkallima
@williamfgc, please check the report at https://app.codecov.io/gh/nkallima/sim-universal-construction (or simply you click at badge).
The coverage is calculated at each push request on v2.5.0-dev.
