### 2021-07-26 9e7512b
Adding documentation for primitives.h. Other minor improvements.
