+++ libconcurrent/includes/primitives.h
@@ -1,3 +1,7 @@
+/// @file primitives.h
+/// @brief This file exposes a simple API for basic atomic operations, basic operations for memory management.
+/// Moreover, this file provides functionality for finding out the the vendor of the processor and some 
+/// basic functionality for measuring time.
 #ifndef _PRIMITIVES_H_
 #define _PRIMITIVES_H_
 
@@ -8,11 +12,17 @@
 #include <stats.h>
 #include <stddef.h>
 
+/// @brief The vendor of the processor is unknown.
 #define UNKNOWN_MACHINE       0x0
+/// @brief The vendor is an AMD x86 processor.
 #define AMD_X86_MACHINE       0x1
+/// @brief The vendor is an Intel X86 processor.
 #define INTEL_X86_MACHINE     0x2
+/// @brief This is a generic X86 processor.
 #define X86_GENERIC_MACHINE   0x3
+/// @brief This is a generic ARM processor.
 #define ARM_GENERIC_MACHINE   0x4
+/// @brief This is a generic RISC-V processor.
 #define RISCV_GENERIC_MACHINE 0x5
 
 #if defined(__GNUC__) && (__GNUC__ * 10000 + __GNUC_MINOR__ * 100) >= 40100
@@ -63,6 +73,7 @@ inline uint64_t nonZeroBits(uint64_t v);
 #endif
 
 #if defined(__GNUC__) && (defined(__amd64__) || defined(__x86_64__))
+/// @brief This macro emits a X86 pause instruction.
 #    define Pause()                         \
         {                                   \
             int __i;                        \
@@ -74,37 +85,140 @@ inline uint64_t nonZeroBits(uint64_t v);
             }                               \
         }
 #else
+/// @brief This macro emits a NO-OP instruction.
 #    define Pause()
 #endif
 
+
+/// @brief This function allocates a memory area of size bytes.
+/// In case that NUMA_SUPPORT is defined in libconcurrent/config.h, the returned memory is allocated on the local NUMA node.
+///
+/// @param size The size of the memory area.
+/// @return In case of error, NULL is returned. In case of success a pointer to the allocated memory area is returned. 
 inline void *getMemory(size_t size);
+
+/// @brief This function allocates a memory area of size bytes. The returned address is aligned to an offset equal to align bytes. 
+/// In case that NUMA_SUPPORT is defined in libconcurrent/config.h, the returned memory is allocated on the local NUMA node.
+///
+/// @param align The alignment size.
+/// @param size The size of the memory area.
+/// @return In case of error, NULL is returned. In case of success a pointer to the allocated memory area is returned.
 inline void *getAlignedMemory(size_t align, size_t size);
+
+/// @brief This function frees memory allocated with either getMemory() or getAlignedMemory() functions.
+///
+/// @param ptr A pointer to the memory area to be freed.
+/// @param size The size of the memory area to be freed.
 inline void freeMemory(void *ptr, size_t size);
+
+/// @brief This function returns the current system's time in milliseconds.
+///
+/// @return System's time in milliseconds.
 inline int64_t getTimeMillis(void);
+
+
+/// @brief This function returns the vendor of the processor that it runs on.
+/// The current version of the Synch framework returns any of the following codes:
+/// - AMD_X86_MACHINE
+/// - INTEL_X86_MACHINE
+/// - X86_GENERIC_MACHINE
+/// - ARM_GENERIC_MACHINE
+/// - RISCV_GENERIC_MACHINE
+/// - UNKNOWN_MACHINE
+///
+/// @return It returns a code for the vendor of the processor that it runs on.  
 inline uint64_t getMachineModel(void);
 
+
+/// A wrapper for the _CAS128 function. See more on _CAS128().
[…]
