### 2021-07-30 93e161a
Improving CONTRIBUTING.md.
