+++ CONTRIBUTING.md
@@ -1,5 +1,74 @@
 # Contribution
 
-If you wish to contribute to the project:
-- Install `clang-format` package (version >= 9)
-- After cloning the repository you need to run `git config --local core.hooksPath .git_hooks` to enable recommended code style checkes (placed in .clang_format file) during git commit. If the tool discovers incosistencies, it will create a patch file. Please follow the instructions to apply the patch before opening a pull request.
+## Source code structure
+The Synch framework consists of 3 main parts, i.e. the Runtime/Primitives, the library of concurrent data-structures and the benchmarks  (see the figure below) . The Runtime/Primitives part provides some basic functionality for creating and managing threads, functionality for basic atomic primitives (e.g. Compare&Swap, Fetch&Add, fences, simple synchronization barriers, etc.), mechanisms for basic memory allocation/management (e.g. memory pools, etc.), functionality for measuring time, reporting CPU counters, etc. Furthermore, the Runtime/Primitives provides a simple and lightweight library of user level-threads that can be used in order to evaluate the provided data-structures and algorithms. 
+
+The Concurrent library utilizes the building blocks of the Runtime/Primitives layer in order to provide all the concurrent data-structures (e.g. combining objects, queues, stacks, etc.). For almost every concurrent data-structure or synchronization mechanism, Synch provides at least one benchmark for evaluating its performance.
+
+<p align="center">
+    <img src="resources/code_structure.png" alt="The code-structure of the Synch framework" width="50%">
+</p>
+
+## First, discuss your issue, change or contribution
+Before contributing to this repository, you are encouraged to first discuss the issue, change or contribution that you wish to make. In order to do this, please open an issue or send an email. We are open to discuss new ideas, improvements or any other contributions. Please note that we have a [Code of conduct](https://github.com/nkallima/sim-universal-construction/blob/main/.github/CODE_OF_CONDUCT.md) that should be considered in all your interactions with the project.
+
+## Ready to make a change
+- Fork the repository.
+- Create a new branch for your patch.
+- Make your updates/changes.
+- Open a pull request.
+- Submit your pull request & get it reviewed.
+
+## Coding conventions
+Before committing a new patch, follow the steps below for checking the coding style of your patch.
+- Ensure that you have installed `clang-format` package (version >= 9).
+- After cloning the repository you need to run `git config --local core.hooksPath .git_hooks` to enable recommended code style checks (placed in `.clang_format` file) during git commit. If the tool discovers inconsistencies, it will create a patch file. Please follow the instructions to apply the patch before opening a pull request.
+
+In general, please conform your coding styling to the following conventions:
+- ident style:
+    * Please avoid using tabs, use spaces instead of tabs.
+- comments:
+    * Comments start with `//`.
+    * For exposing documentation to Doxygen start comments with `///`.
+- structs:
+    * Self-explanatory names: a struct `DoubleEndedQueueStruct` will tell the developer what the struct is used for.
+    * Use a capital-letter for the first character of the struct.
+    * Please use `typedef` for simplifying struct-naming, e.g. `DoubleEndedQueueStruct` instead of `struct DoubleEndedQueueStruct`.
+- functions: for functions the following rules hold:
+    * Self-explanatory names: a function `getMemory()` will tell the developer what it returns as well as `getThreadId()`, etc.
+    * Avoid using a capital-letter for the first character of the function.
+    * All the public functions provided to the end-user should start with the `synch` prefix (starting from v.3.0.0), i.e. `synchGetMemory`.
+    * All the internal functions should NOT start with the `synch` prefix, i.e. `getMemory`.
+- memory allocation and alignment:
+    * For memory management, please do not use `malloc`, `calloc`, etc. functions directly.
[…]
