### 2021-07-30 nkallima/sim-universal-construction#34 @nkallima
@williamfgc
I have added some new material in the Contributing documentation (see https://github.com/nkallima/sim-universal-construction/blob/v2.5.0-dev/CONTRIBUTING.md).
