### 2021-08-02 #3143 @williamfgc
> @williamfgc - any updates?
@danielskatz yes, I've been working with @nkallima on addressing some of the issues I opened. Thanks @nkallima for addressing those, it's looking good.

@nkallima I have a couple of comments on the manuscript related to the JOSS review process:

Summary:
- **Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?**:  The Summary section does a good job in explaining some of the computational aspects. Please list some of the merits of the software itself in that section to motivate why first time user should use or contribute to the software. The narrative is up to you, but as an example I'd mention something along the lines of: "Synch is a C library, thoroughly tested targeting x86_64 POSIX systems. We provide comprehensive documentation for the first-time. Developers will find a comprehensive set of tests to ensure quality and reproducibility of the results."

-  **State of the field: Do the authors describe how this software compares to other commonly-used packages?** Are there any other similar efforts in terms of software functionality and framework? Adding some background information would be good to understand what the unique functionalities of Synch are for those familiar with other tools.

I hope it helps, I don't think I have any other major points from my side. I haven't been able to verify benchmarking claims as I'm not aware or familiar with standard benchmarks or points of reference in this field. Tests run fine, though.

Minor recommendations (not necessarily required):

- [ ] Add a statement to indicate what the benchmark example goal is [ReadMe](https://github.com/nkallima/sim-universal-construction#api---code-example-for-a-simple-benchmark). While the section describes what the bechmarks does, it's not clear if the user is supposed to look at weak or strong scalability.

- [ ] While the code base is straight-forward. Add CI and containers (e.g. docker) to enable portability and reproducibility of the results and as I means for quick sharing.
