### 2021-08-02 nkallima/sim-universal-construction#34 @nkallima
Thanks @williamfgc for your reply & your comments. I improved this point in the last commit (https://github.com/nkallima/sim-universal-construction/commit/50daa36ab51c0ecf2dfe29fc2fb8528652ade814).
In case you think that the Contributing file is in a good state, this issue could be closed.

>
>
> Thanks @nkallima looking good. Minor thing, please indicate in [this section](https://github.com/nkallima/sim-universal-construction/blob/v2.5.0-dev/CONTRIBUTING.md#ready-to-make-a-change) the target branch (is it `main`?) for the pull request. Otherwise, looks good.
