### 2021-08-15 #3143 @jarrah42
@nkallima perhaps I missed it. The documentation only mentions x86_64 but does not mention that it also depends on Linux. I tried building on Darwin but it doesn't compile, so you should list the O/S dependency also.
