### 2021-08-16 #3143 @danielskatz
We'll make sure the version number is set correctly at the end of the process, in case it changes again. This is part of the publishing step, along with creating an archive of the repo at that point
