### 2021-08-18 #3143 @danielskatz
@williamfgc and @nkallima - how is the resolution of the remaining 4 items on @williamfgc's checklist coming?
