### 2021-08-18 9c3434c
Providing a short discussion (PERFORMANCE.md) of the expected performance.
