+++ PERFORMANCE.md
@@ -1,0 +1,67 @@
+# Expected Performance
+
+This document provides a brief discussion of the expected performance for some concurrent objects provided by the Synch framework. For a more detailed performance analysis of the provided objects, the reader is encouraged to look at [1,2,3,10].
+
+## Evaluation testbed
+
+The provided brief performance analysis is performed  in a 64-core AMD Epyc multiprocessor (abbr. Epyc) consisting of 2 Epyc 7501 processors that is able to handle 128 threads. Each of these processors consists of 4 silicon dies and each silicon die consist of a NUMA node. Each die contains $8$ processing cores, where each core is handling 2 threads. Thus, the AMD multiprocessor consists of 8 NUMA nodes, where each NUMA node consists of 8 processing cores.
+
+## Combining objects performance
+
+At the first experiment, the CC-Synch [1], DSM-Synch [1], H-Synch [1] and PSim [2] combining objects are compared with a universal object based on MCS spin-locks [9] (provided by the `mcsbench.run` benchmark), a simple lock-free implementation (provided by the `lfuobjectbench.run` benchmark), and the flat-combining (abbr. FC) technique provided by [11, 12].
+
+In the experiment of the following figure, the performance of the evaluated combining objects is presented while performing the Fetch&Multiply synthetic benchmark. In this benchmark, a simple concurrent object that supports an atomic Fetch&Multiply float object is implemented. This object is simple enough to exhibit any overheads that a synchronization technique may induce while simulating a very simple concurrent object under high contention. The figure shows the average throughput (Fetch&Multiply operations per second) that each synchronization technique achieves when it executes $10^7$ Fetch&Multiply operations (i.e., by setting `-r=10000000`). Since for different values of threads $t$, each thread executes  $10^7/t$ Fetch&Multiply operations, this experiment measures strong scaling. Specifically, the horizontal axis represents the number of threads $t$, while the vertical axis displays the throughput (in millions of operations per second) that each synchronization technique achieves. For each value of $t$, the experiment has been performed 10 times and averages have been calculated using the `bench.sh` script. A random number of dummy loop iterations (i.e, by setting `-w=512`) have been inserted between the execution of two consecutive Fetch&Multiply operations by each thread. In this way, a random work load large enough to avoid unrealistically low cache miss ratios and long runs is simulated. However, this workload is not big enough to substantially reduce the contention. In the experiments performed on the Epyc machine, H-Synch outperforms all other synchronization techniques since it exploits the hierarchical NUMA nature of the Epyc machine. More specifically, H-Synch achieves up to $3$ higher throughput than flat-combining and outperforms CC-Synch by a factor of up to $1.23$. The simple lock-free implementation of Fetch&Multiply is slightly slower than PSim and faster that flat-combining. Also, CC-Synch is up to $6$ times faster than the object based on MCS spin-locks. CC-Synch performs also very well; its performance is close to that of DSM-Synch and in par with the lock-free implementation.
+
+<p align="center">
+    <img src="resources/fam_epyc_w512.png" width="80%">
+</p>
+
[…]
