### 2021-08-18 nkallima/sim-universal-construction#31 @nkallima
@williamfgc

>  What's the baseline used for performance claims? Is it measuring strong/weak scalability?
>  Illustrate examples comparing this implementation with existing implementations to justify performance (e.g. C++ std::queue, std::unordered_map for hash-table, C++ etc?) to users know what to expect in terms of performance.

I have added a new document that describes the expected performance of various objects/data-structures provided by the framework. It also provides some indicative performance measurements. Please check  https://github.com/nkallima/sim-universal-construction/blob/v2.5.0-dev/PERFORMANCE.md.
Moreover, README.md and PERFORMANCE.md clearly state that strong scaling is measured.
