### 2021-08-18 #3143 @nkallima
@williamfgc @danielskatz

I was working towards closing the issue https://github.com/nkallima/sim-universal-construction/issues/31.
Specifically, I have added a new document that describes the expected performance of various objects/data-structures provided by the framework. I also provide some indicative performance measurements. Please check https://github.com/nkallima/sim-universal-construction/blob/v2.5.0-dev/PERFORMANCE.md.

Now, it remains to slightly modify the naming of some parts of the API as was suggested by williamfgc in order to close the issue. Next, I will update the text of the manuscript following the suggestions of williamfgc.
