### 2021-08-18 e6fdf1e
Update README.md and PERFORMANCE.md.
