### 2021-08-19 nkallima/sim-universal-construction#31 @nkallima
@williamfgc

> Indicate if this is a library, executable. If it's a C library as it seems to be the case from the API example, please add a prefix_ "namespace" (e.g. synch_foo) to library API functions to avoid name clashes

I have added a `synch` prefix for the provided functions of the public API. I also added the `Synch` prefix for the public structs. A first version of this new naming scheme can be found in the v3.0.0-dev branch  https://github.com/nkallima/sim-universal-construction/tree/v3.0.0-dev. Given that the new version of the API is incompatible with the previous versions, I had to increase the major number of the version to 3.
However, this naming scheme introduces the following issues:

* The functions’ naming for some of the provided objects are weird. For example, the `CCSynchStructInit` function of the CC-Synch object converts to `synchCCSynchStructInit`, which is a bit weird since there are two instances of the `synch`. The same holds for a few other objects such as H-Synch, DSM-Synch, etc.
 * It creates confusion to some cases. For example the `SimStructInit` function for the Sim combining object converts to `synchSimStructInit`. This may confuse a user for which object the `synchSimStructInit` function stands for, Sim or  CC-Synch? This holds for a lot of objects.
For these problems, I thought of the following three solutions:
1. Leaving the API as is in the v3.0.0-dev branch.
2. For all functions for the runtime part of the framework (i.e, functions for creating threads, functions for managing memory, etc.), the `synch` prefix could be adopted. The legacy naming (without the `synch` prefix) could be used only for the concurrent objects, e.g. CC-Synch, Sim, etc.
3. Keeping the API of v2.x.y versions.

In my opinion, 2 seems to be a balanced trade-off. What do you think?
