### 2021-08-19 4e0e259
Adding the `Synch` prefix to all structs provided by primitives/runtime to the end-user. Adding the `synch` prefix to all functions provided by primitives/runtime to the end-user.
