+++ libconcurrent/concurrent/osciqueue.c
@@ -4,11 +4,10 @@
 inline static RetVal serialEnqueue(void *state, ArgVal arg, int pid);
 inline static RetVal serialDequeue(void *state, ArgVal arg, int pid);
 
-
 void OsciQueueInit(OsciQueueStruct *queue_object_struct, uint32_t nthreads, uint32_t fibers_per_thread) {
     OsciInit(&(queue_object_struct->enqueue_struct), nthreads, fibers_per_thread);
     OsciInit(&queue_object_struct->dequeue_struct, nthreads, fibers_per_thread);
-    queue_object_struct->pool_node = getAlignedMemory(CACHE_LINE_SIZE, queue_object_struct->enqueue_struct.groups_of_fibers * sizeof(PoolStruct));
+    queue_object_struct->pool_node = synchGetAlignedMemory(CACHE_LINE_SIZE, queue_object_struct->enqueue_struct.groups_of_fibers * sizeof(SynchPoolStruct));
     queue_object_struct->guard.val = GUARD_VALUE;
     queue_object_struct->guard.next = NULL;
     queue_object_struct->first = &queue_object_struct->guard;
@@ -18,14 +17,14 @@ void OsciQueueInit(OsciQueueStruct *queue_object_struct, uint32_t nthreads, uint
 void OsciQueueThreadStateInit(OsciQueueStruct *object_struct, OsciQueueThreadState *lobject_struct, int pid) {
     OsciThreadStateInit(&lobject_struct->enqueue_thread_state, &object_struct->enqueue_struct, (int)pid);
     OsciThreadStateInit(&lobject_struct->dequeue_thread_state, &object_struct->dequeue_struct, (int)pid);
-    init_pool(&(object_struct->pool_node[getThreadId()]), sizeof(Node));
+    synchInitPool(&(object_struct->pool_node[synchGetThreadId()]), sizeof(Node));
 }
 
 inline static RetVal serialEnqueue(void *state, ArgVal arg, int pid) {
     OsciQueueStruct *st = (OsciQueueStruct *)state;
     Node *node;
 
-    node = alloc_obj(&(st->pool_node[getThreadId()]));
+    node = synchAllocObj(&(st->pool_node[synchGetThreadId()]));
     node->next = NULL;
     node->val = arg;
     st->last->next = node;
@@ -43,7 +42,7 @@ inline static RetVal serialDequeue(void *state, ArgVal arg, int pid) {
         node = st->first;
         if (node->val == GUARD_VALUE)
             return serialDequeue(state, arg, pid);
-        recycle_obj(&(st->pool_node[getThreadId()]), (Node *)prev);
+        synchRecycleObj(&(st->pool_node[synchGetThreadId()]), (Node *)prev);
         return node->val;
     } else {
         return EMPTY_QUEUE;
