+++ benchmarks/msqueuebench.c
@@ -16,8 +16,8 @@
 MSQueueStruct queue CACHE_ALIGN;
 int64_t d1 CACHE_ALIGN, d2;
 int MIN_BAK, MAX_BAK;
-Barrier bar CACHE_ALIGN;
-BenchArgs bench_args CACHE_ALIGN;
+SynchBarrier bar CACHE_ALIGN;
+SynchBenchArgs bench_args CACHE_ALIGN;
 
 inline static void *Execute(void *Arg) {
     MSQueueThreadState *th_state;
@@ -26,39 +26,38 @@ inline static void *Execute(void *Arg) {
     long rnum;
     volatile long j;
 
-    th_state = getAlignedMemory(CACHE_LINE_SIZE, sizeof(MSQueueThreadState));
+    th_state = synchGetAlignedMemory(CACHE_LINE_SIZE, sizeof(MSQueueThreadState));
     MSQueueThreadStateInit(th_state, bench_args.backoff_low, bench_args.backoff_high);
-    fastRandomSetSeed(id + 1);
-    BarrierWait(&bar);
-    if (id == 0)
-        d1 = getTimeMillis();
+    synchFastRandomSetSeed(id + 1);
+    synchBarrierWait(&bar);
+    if (id == 0) d1 = synchGetTimeMillis();
 
     for (i = 0; i < bench_args.runs; i++) {
         MSQueueEnqueue(&queue, th_state, id);
-        rnum = fastRandomRange(1, bench_args.max_work);
+        rnum = synchFastRandomRange(1, bench_args.max_work);
         for (j = 0; j < rnum; j++)
             ;
         MSQueueDequeue(&queue, th_state);
-        rnum = fastRandomRange(1, bench_args.max_work);
+        rnum = synchFastRandomRange(1, bench_args.max_work);
         for (j = 0; j < rnum; j++)
             ;
     }
-    BarrierWait(&bar);
-    if (id == 0) d2 = getTimeMillis();
+    synchBarrierWait(&bar);
+    if (id == 0) d2 = synchGetTimeMillis();
 
     return NULL;
 }
 
 int main(int argc, char *argv[]) {
-    parseArguments(&bench_args, argc, argv);
+    synchParseArguments(&bench_args, argc, argv);
 
     MSQueueInit(&queue);
-    BarrierSet(&bar, bench_args.nthreads);
-    StartThreadsN(bench_args.nthreads, Execute, bench_args.fibers_per_thread);
-    JoinThreadsN(bench_args.nthreads - 1);
+    synchBarrierSet(&bar, bench_args.nthreads);
+    synchStartThreadsN(bench_args.nthreads, Execute, bench_args.fibers_per_thread);
+    synchJoinThreadsN(bench_args.nthreads - 1);
 
     printf("time: %d (ms)\tthroughput: %.2f (millions ops/sec)\t", (int)(d2 - d1), 2 * bench_args.runs * bench_args.nthreads / (1000.0 * (d2 - d1)));
-    printStats(bench_args.nthreads, bench_args.total_runs);
+    synchPrintStats(bench_args.nthreads, bench_args.total_runs);
 #ifdef DEBUG
     long counter = 0;
 
