### 2021-08-19 947afd5
Adding the `synch` prefix in the functionality provided by primitives.h.
