+++ libconcurrent/concurrent/lcrq.c
@@ -104,14 +104,13 @@ inline static int crq_is_closed(uint64_t t) {
 
 inline static void fix_state(RingQueue *rq) {
     while (true) {
-        uint64_t t = FAA64(&rq->tail, 0);
-        uint64_t h = FAA64(&rq->head, 0);
+        uint64_t t = synchFAA64(&rq->tail, 0);
+        uint64_t h = synchFAA64(&rq->head, 0);
 
-        if (Unlikely(rq->tail != t))
-            continue;
+        if (synchUnlikely(rq->tail != t)) continue;
 
         if (h > t) {
-            if (CAS64(&rq->tail, t, h)) break;
+            if (synchCAS64(&rq->tail, t, h)) break;
             continue;
         }
         break;
@@ -120,9 +119,9 @@ inline static void fix_state(RingQueue *rq) {
 
 inline static int close_crq(RingQueue *rq, const uint64_t t, const int tries) {
     if (tries < 10)
-        return CAS64(&rq->tail, t + 1, set_unsafe(t + 1));
+        return synchCAS64(&rq->tail, t + 1, set_unsafe(t + 1));
     else
-        return BitTAS64(&rq->tail, 63);
+        return synchBitTAS64(&rq->tail, 63);
 }
 
 void LCRQInit(LCRQStruct *queue, uint32_t nthreads UNUSED_ARG) {
@@ -145,19 +144,18 @@ void LCRQEnqueue(LCRQStruct *queue, LCRQThreadState *thread_state, ArgVal arg, i
         RingQueue *rq = queue->tail;
 
 #ifdef HAVE_HPTRS
-        SWAP(&thread_state->hazardptr, rq);
-        if (Unlikely(queue->tail != rq))
-            continue;
+        synchSWAP(&thread_state->hazardptr, rq);
+        if (synchUnlikely(queue->tail != rq)) continue;
 #endif
 
         RingQueue *next = rq->next;
 
-        if (Unlikely(next != NULL)) {
-            CASPTR(&queue->tail, rq, next);
+        if (synchUnlikely(next != NULL)) {
+            synchCASPTR(&queue->tail, rq, next);
             continue;
         }
 
-        uint64_t t = FAA64(&rq->tail, 1);
+        uint64_t t = synchFAA64(&rq->tail, 1);
 
         if (crq_is_closed(t)) {
 alloc:
@@ -169,8 +167,8 @@ alloc:
             // Solo enqueue
             thread_state->nrq->tail = 1, thread_state->nrq->array[0].val = arg, thread_state->nrq->array[0].idx = 0;
 
-            if (CASPTR(&rq->next, NULL, thread_state->nrq)) {
-                CASPTR(&queue->tail, rq, thread_state->nrq);
+            if (synchCASPTR(&rq->next, NULL, thread_state->nrq)) {
+                synchCASPTR(&queue->tail, rq, thread_state->nrq);
                 thread_state->nrq = NULL;
                 return;
             }
@@ -178,22 +176,22 @@ alloc:
         }
 
         RingNode *cell = &rq->array[t & (RING_SIZE - 1)];
-        StorePrefetch(cell);
+        synchStorePrefetch(cell);
 
         uint64_t idx = cell->idx;
         uint64_t val = cell->val;
 
-        if (Likely(is_empty(val))) {
-            if (Likely(node_index(idx) <= t)) {
-                if ((Likely(!node_unsafe(idx)) || rq->head < t) && CAS128((uint64_t*)cell, -1, idx, arg, t)) {
+        if (synchLikely(is_empty(val))) {
+            if (synchLikely(node_index(idx) <= t)) {
+                if ((synchLikely(!node_unsafe(idx)) || rq->head < t) && synchCAS128((uint64_t *)cell, -1, idx, arg, t)) {
                     return;
                 }
[…]
