+++ libconcurrent/concurrent/dsmqueue.c
@@ -30,7 +30,7 @@ inline static RetVal serialEnqueue(void *state, ArgVal arg, int pid) {
     node->val = arg;
     st->last->next = node;
     st->last = node;
-    NonTSOFence();
+    synchNonTSOFence();
     return ENQUEUE_SUCCESS;
 }
 
@@ -44,7 +44,7 @@ inline static RetVal serialDequeue(void *state, ArgVal arg, int pid) {
         node = st->first;
         if (node->val == GUARD_VALUE)
             return serialDequeue(state, arg, pid);
-        NonTSOFence();
+        synchNonTSOFence();
         synchRecycleObj(&pool_node, (Node *)prev);
         return node->val;
     } else {
