### 2021-08-19 c80887d
A fix for include files in the installation script.
