### 2021-08-19 nkallima/sim-universal-construction#31 @nkallima
@williamfgc thanks for your feedback! The revised API is now available at v.3.0.0-dev branch (https://github.com/nkallima/sim-universal-construction/tree/v3.0.0-dev).
In case that you are OK with the changes, we could close this issue.
