+++ CHANGELOG.md
@@ -7,7 +7,7 @@ v3.0.0
 - Adding the `Synch` prefix to all structs provided by the runtime/primitives to the end-user.
 - Adding the `synch` prefix to all functions provided by the runtime/primitives to the end-user.
 - Extensive documentation for the API.
-- Create a `doxygen` configuration file dor auto-generating man-pages/documentation.
+- Create a `doxygen` configuration file for auto-generating man-pages/documentation.
 - API documentation is now provided at GitHub pages: https://nkallima.github.io/sim-universal-construction/index.html.
 - A code coverage report is provided through codecov.io for the validation script.
 - Support for installing the framework.
