+++ paper.bib
@@ -167,3 +167,66 @@
   number={RJ-5118},
   institution={International Business Machines Incorporated, Thomas J. Watson Research}
 }
+
+@inproceedings{LCRQ,
+  author = {Morrison, Adam and Afek, Yehuda},
+  title = {Fast Concurrent Queues for X86 Processors},
+  year = {2013},
+  isbn = {9781450319225},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  url = {https://doi.org/10.1145/2442516.2442527},
+  doi = {10.1145/2442516.2442527},
+  booktitle = {Proceedings of the 18th ACM SIGPLAN Symposium on Principles and Practice of Parallel Programming},
+  pages = {103–112},
+  numpages = {10},
+  keywords = {concurrent queue, fetch-and-add, nonblocking algorithm},
+  location = {Shenzhen, China},
+  series = {PPoPP '13}
+}
+
+@electronic{LCRQurl,
+ author = {Morrison, Adam and Afek, Yehuda},
+ title = {Code for LCRQ},
+ url = {http://mcg.cs.tau.ac.il/projects/lcrq}
+}
+
+@InProceedings{scal,
+  author={Haas, Andreas and H{\"u}tter, Thomas and Kirsch, Christoph M. and Lippautz, Michael and Preishuber, Mario and Sokolova, Ana},
+  editor={Bouajjani, Ahmed and Fauconnier, Hugues},
+  title={Scal: A Benchmarking Suite for Concurrent Data Structures},
+  doi={10.1007/978-3-319-26850-7_1},
+  booktitle={Networked Systems},
+  year={2015},
+  publisher={Springer International Publishing},
+  pages={1-14},
+  isbn="978-3-319-26850-7"
+}
+
+@inproceedings{FC,
+  author = {Hendler, Danny and Incze, Itai and Shavit, Nir and Tzafrir, Moran},
+  title = {Flat Combining and the Synchronization-Parallelism Tradeoff},
+  year = {2010},
+  isbn = {9781450300797},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  url = {https://doi.org/10.1145/1810479.1810540},
+  doi = {10.1145/1810479.1810540},
+  booktitle = {Proceedings of the Twenty-Second Annual ACM Symposium on Parallelism in Algorithms and Architectures},
+  pages = {355–364},
+  numpages = {10},
+  keywords = {concurrent data-structures, synchronization, multiprocessors},
+  location = {Thira, Santorini, Greece},
+  series = {SPAA '10}
+}
+
+@electronic{FCurl,
+ author = {Hendler, Danny and Incze, Itai and Shavit, Nir and Tzafrir, Moran},
+ title = {Code for flat-combining},
+ url = {https://github.com/mit-carbon/Flat-Combining}
+}
+
+@electronic{CDSurl,
+ title = {The Concurrent Data Structures (CDS) library},
+ url = {https://github.com/khizmax/libcds}
+}
