### 2021-08-20 5895ab6
Adding the `SYNCH_` prefix to most of the public definitions.
