+++ libconcurrent/libdeps.sh
@@ -2,7 +2,7 @@
 
 LDLIBS="-lpthread -latomic";
 
-DEFINITIONS=(NUMA_SUPPORT _TRACK_CPU_COUNTERS);
+DEFINITIONS=(SYNCH_NUMA_SUPPORT SYNCH_TRACK_CPU_COUNTERS);
 LIBS=("-lnuma" "-lpapi");
 
 for i in ${!DEFINITIONS[@]}; do
