+++ libconcurrent/primitives/pool.c
@@ -75,7 +75,7 @@ void *synchAllocObj(SynchPoolStruct *pool) {
 }
 
 void synchRecycleObj(SynchPoolStruct *pool, void *obj) {
-#ifndef POOL_NODE_RECYCLING_DISABLE
+#ifndef SYNCH_POOL_NODE_RECYCLING_DISABLE
     SynchBlockObject *object = obj;
     object->next = pool->recycle_list;
     pool->recycle_list = object;
