### 2021-08-20 41ffa39
Supressing a warning and minor fix in memory allocation in AMD machines for DSM-Synch.
