+++ libconcurrent/primitives/primitives.c
@@ -3,6 +3,7 @@
 #include <malloc.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 
 #ifdef SYNCH_NUMA_SUPPORT
 #    include <numa.h>
