+++ paper.bib
@@ -230,3 +230,8 @@
  title = {The Concurrent Data Structures (CDS) library},
  url = {https://github.com/khizmax/libcds}
 }
+
+@electronic{Boost,
+ title = {BOOST C++ libraries},
+ url = {www.boost.org}
+}
