### 2021-08-20 #3143 @nkallima
@williamfgc

Thanks for your comments and feedback!

> * **Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?**:  The Summary section does a good job in explaining some of the computational aspects. Please list some of the merits of the software itself in that section to motivate why first time user should use or contribute to the software. The narrative is up to you, but as an example I'd mention something along the lines of: "Synch is a C library, thoroughly tested targeting x86_64 POSIX systems. We provide comprehensive documentation for the first-time. Developers will find a comprehensive set of tests to ensure quality and reproducibility of the results."

I have added some text addressing this.

> * **State of the field: Do the authors describe how this software compares to other commonly-used packages?** Are there any other similar efforts in terms of software functionality and framework? Adding some background information would be good to understand what the unique functionalities of Synch are for those familiar with other tools.

I have added a related work section that describes similar packages.

> * [ ]  Add a statement to indicate what the benchmark example goal is [ReadMe](https://github.com/nkallima/sim-universal-construction#api---code-example-for-a-simple-benchmark). While the section describes what the bechmarks does, it's not clear if the user is supposed to look at weak or strong scalability.

I have added some extra material for handling this.

@danielskatz, @williamfgc, @jarrah42

Following the advice of  williamfgc, I have made some modifications on the naming of the API. Since the revised API is incompatible with the older version,s the version of the code should increase to v3.0.0. All improvements on the repository that are made during the review and revised API are now available at v.3.0.0-dev branch (https://github.com/nkallima/sim-universal-construction/tree/v3.0.0-dev).

@danielskatz
All issues of the repository are closed and I have applied all the requested changes to the text so far.
