### 2021-08-20 2a98557
Improving a bit the `validation.sh` script for codecov.
