### 2021-08-20 b0ae65a
The `bench.sh` and `validate.sh` scripts use similar arguments as the provided benchmarks.
