+++ validate.sh
@@ -15,7 +15,7 @@ PASS_STATUS=1
 
 function usage()
 {
-    echo -e "Usage: ./validate.sh OPTION=NUM ...";
+    echo -e "Usage: ./validate.sh OPTION1 VALUE1 OPTION2 VALUE2 ...";
     echo -e "This script compiles the sources in DEBUG mode and validates the correctnes of some of the provided concurrent objects.";
     echo -e ""
     echo -e "The following options are available."
@@ -43,8 +43,9 @@ if [ "$1" = "--help" ] || [ "$1" = "-h" ]; then
 fi
 
 while [ "$1" != "" ]; do
-    PARAM=`echo $1 | awk -F= '{print $1}'`
-    VALUE=`echo $1 | awk -F= '{print $2}'`
+    PARAM=`echo $1`
+    VALUE=`echo $2`
+    SHIFT=0
     case $PARAM in
         -h | --help)
             usage;
@@ -53,21 +54,27 @@ while [ "$1" != "" ]; do
         -s | --step_threads)
             STEP_THREADS=$VALUE
             STEP_SELETCTED=1
+            SHIFT=1
             ;;
         -t | --max_threads)
             MAX_PTHREADS=$VALUE
+            SHIFT=1
             ;;
         -f | --fibers)
             FIBERS="-f $VALUE"
+            SHIFT=1
             ;;
         -w | --max_work)
             WORKLOAD="-w $VALUE"
+            SHIFT=1
             ;;
         -n | --numa_nodes)
             NUMA_NODES="-n $VALUE"
+            SHIFT=1
             ;;
         -r | --runs)
             RUNS_PER_THREAD=$VALUE
+            SHIFT=1
             ;;
         --codecov)
             CODECOV=1
@@ -77,11 +84,16 @@ while [ "$1" != "" ]; do
             usage
             exit 1
             ;;
-	*)
-	    FILE=$PARAM
-	    ;;
     esac
     shift
+    if [ $SHIFT = "1" ]; then
+        shift
+        if [ "$VALUE" = "" ]; then
+            echo "ERROR: no value set for \"$PARAM\""
+            usage
+            exit 1
+        fi
+    fi
 done
 
 # Calculate step according to user preferences
