### 2021-08-24 3f0bff9
Improving script for deploying documentation.
