### 2021-08-31 #3143 @nkallima
@danielskatz

> My suggested changes from proofreading are in [nkallima/sim-universal-construction#39](https://github.com/nkallima/sim-universal-construction/pull/39)

Thank you for leading the review and for your suggestions! I applied your patch in `joss-branch`.

> @nkallima - the next step will for me to proofread the paper, and then for you to:
>
>     * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
>
>     * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
>
>     * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>
>     * [ ]  Please list the DOI of the archived version here.
>
>
> I'll let you know when I'm done proof-reading.

- A tagged release of the software was made. Version v3.0.0
- Software DOI: 10.5281/zenodo.5347958
