### 2021-08-31 32aec56
fixes to things I did wrong in paper
