### 2021-08-31 #3143 @danielskatz
@nkallima - Unfortunately, a couple of the changes I made to expose references as text (rather than in parens) missed changing the required ";"s to better things in text. Merging https://github.com/nkallima/sim-universal-construction/pull/44 should fix these errors I made.
